% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_operations.R
\name{add_assay}
\alias{add_assay}
\title{Adds an assay from an \code{\link{scMethrix}} object}
\usage{
add_assay(scm = NULL, new_assay = "new_assay", matrix = NULL)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{new_assay}{string; name for transformed assay. Default = "new_assay"}

\item{matrix}{mtx; the input matrix}
}
\value{
An \code{\link{scMethrix}} object
}
\description{
Adds an assay from an \code{\link{scMethrix}} object
}
\details{
Simple
Fulfills the same function as \code{assay(scm, assay) <- matrix}, but with additional checks.#'
}
\examples{
data('scMethrix_data')
}
