% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_clustering.R
\name{append_col_data}
\alias{append_col_data}
\title{Appends colData in an scMethrix object}
\usage{
append_col_data(scm, colData)
}
\arguments{
\item{scm}{scMethrix; Input \code{\link{scMethrix}} object}

\item{colData}{matrix; A matrix containing colData. Must contain either a column labelled "Sample" or row names that correspond with the input \code{\link{scMethrix}} object}
}
\value{
An \code{\link{scMethrix}} object
}
\description{
Appends colData in an scMethrix object
}
\details{
Typically used for clustering. Allows additional information to be added to colData in an scMethrix object after the object creation. It does this via a left join on the original colData. Any samples not included in the colData object will be filled with NAs.
}
\examples{
data('scMethrix_data')
colData <- colData(scMethrix_data)
colData["Type"] <- "Cell"
scMethrix_data <- append_col_data(scMethrix_data,colData)
colData(scMethrix_data)
}
