% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_funcs.R
\name{arg.match}
\alias{arg.match}
\title{Allows partial pattern matching in a responsive manner}
\usage{
arg.match(parent, arg)
}
\arguments{
\item{parent}{closure; the parent function in which to check the input}

\item{arg}{varibale; the variable in which to check}
}
\value{
arg, if the value is in the function definition.
}
\description{
Allows partial pattern matching in a responsive manner
}
\details{
Check the parent function input arguments to see whether the inputted value is part of the set. Will return a formatted error message with the incorrect variable name and all the acceptable inputs
}
