% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_object.R
\name{as.scMethrix.GRset}
\alias{as.scMethrix.GRset}
\title{Converts from a minfi::GRset to an scMethrix object}
\usage{
as.scMethrix.GRset(GRset, colData = NULL)
}
\arguments{
\item{GRset}{GRset; the GRset to convert to scMethrix}

\item{colData}{data.table; information about the samples}
}
\value{
An \code{\link{scMethrix}} object
}
\description{
Converts from a minfi::GRset to an scMethrix object
}
\examples{
data('scMethrix_data')
}
