% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_funcs.R
\name{bin_granges}
\alias{bin_granges}
\title{Bins a \code{\link{GRanges}} object into bins of length \code{bin_size}}
\usage{
bin_granges(gr, bin_size = 1e+05)
}
\arguments{
\item{gr}{The \code{\link{GRanges}} object}

\item{bin_size}{The size of bin to use}
}
\value{
\code{\link{GRangesList}} containing all the binned \code{\link{GRanges}}
}
\description{
Bins a \code{\link{GRanges}} object into bins of length \code{bin_size}
}
\details{
Bins \code{\link{GRanges}} objects into a specified bin size by chromosome
}
\examples{
regions <- GenomicRanges::GRanges(seqnames = "chr1", ranges = IRanges(1,10000))
bin_granges(regions,bin_size=1000) 
}
