% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_transforms.R
\name{bin_scMethrix}
\alias{bin_scMethrix}
\title{Bins the ranges of an \code{\link{scMethrix}} object.}
\usage{
bin_scMethrix(scm, bin_size = 1e+05, trans = NULL, h5_dir = NULL)
}
\arguments{
\item{scm}{A \code{\link{scMethrix}} object}

\item{bin_size}{The size of each bin. First bin will begin at the start position of the first genomic
region on the chromosome}

\item{trans}{The transforms for each assay. Must be a named vector of functions (closure).
Default = mean(.., na.rm=TRUE)}

\item{h5_dir}{directory to store H5 based object}
}
\value{
An \code{\link{scMethrix}} object
}
\description{
Bins the ranges of an \code{\link{scMethrix}} object.
}
\details{
Uses the inputted function to transform an assay in the \code{\link{scMethrix}} object
}
\examples{
data('scMethrix_data')
trans <- c(score = function(x) mean(x,na.rm=TRUE),counts = function(x) sum(x,na.rm=TRUE))
bin_scMethrix(scMethrix_data,trans = trans)
}
