% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_funcs.R
\name{binarize}
\alias{binarize}
\title{Binarize an input value based on a \code{threshold}}
\usage{
binarize(x, threshold = NULL, rep_na = NA, verbose = FALSE)
}
\arguments{
\item{x}{A value to binarize}

\item{threshold}{The threshold for binarizing. Will default to the center number between max and min.}

\item{rep_na}{The value to replace missing values with. Default NA.}

\item{verbose}{boolean; flag for whether to display threshold or not}
}
\value{
1 or 0, if above of below the threshold, or 'rep.na' if NA
}
\description{
Binarize an input value based on a \code{threshold}
}
\details{
Assigns a value of 0 or 1 based on being < or > the \code{thresdhold}, respectively.
If \code{x} = \code{threshold}, \code{x} = 0. NA values are assigned as -1
}
\examples{
vals <- c(0,0.25,0.5,0.75,1,NA)
binarize(vals, threshold=0.5)
}
