% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_ops.R
\name{cast_datatable}
\alias{cast_datatable}
\title{Casts genomic regions into \code{\link{data.table}} format}
\usage{
cast_datatable(regions)
}
\arguments{
\item{regions}{GRanges or data.frame; The input regions to cast to \code{\link{GRanges}}}
}
\value{
\code{\link{data.table}} object with the input regions
}
\description{
Casts genomic regions into \code{\link{data.table}} format
}
\details{
Casts the input as a \code{\link{data.table}} object. Input can be \code{\link{GRanges}} or a
\code{\link{data.frame}}-compatible class that can be cast through \code{as.data.frame()}. Input BED format
must be \code{chr-start-end} for \code{\link{data.frame}} objects.
}
\examples{
regions = data.table(chr = 'chr1', start = 1, end = 100)
cast_granges(regions) 
}
