% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_funcs.R
\name{cast_granges}
\alias{cast_granges}
\title{Casts regions into Granges format}
\usage{
cast_granges(regions)
}
\arguments{
\item{regions}{The input regions}
}
\value{
GRanges object with the input regions
}
\description{
Casts regions into Granges format
}
\details{
Casts the input as a GRanges object. Input can be GRange or a data.frame-compatible class that
can be cast through as.data.frame(). Input format must be chr-start-end for data.frame objects.
}
\examples{
regions = data.table(chr = 'chr1', start = 1, end = 100)
cast_granges(regions) 
}
