% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_funcs.R
\name{cbindlist}
\alias{cbindlist}
\title{A faster version of cbind when trying to combine lists of data.tables}
\usage{
cbindlist(list)
}
\arguments{
\item{list}{A list of data.tables with identical # of rows}
}
\value{
data.table; the cbinded output
}
\description{
A faster version of cbind when trying to combine lists of data.tables
}
