% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_funcs.R
\name{cd}
\alias{cd}
\title{Same as colData(scm), but shorter syntax, and will output row names if there is no columns}
\usage{
cd(scm)
}
\arguments{
\item{scm}{an scMethrix}
}
\description{
Same as colData(scm), but shorter syntax, and will output row names if there is no columns
}
