% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_funcs.R
\name{check.scm}
\alias{check.scm}
\title{Checks to see if object is a proper scMethrix object}
\usage{
check.scm(scm)
}
\arguments{
\item{scm}{scMethrix; the experiment object to test}
}
\value{
invisible(TRUE), if the object is valid. Error if not.
}
\description{
Checks to see if object is a proper scMethrix object
}
