% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_clustering.R
\name{cluster_scMethrix}
\alias{cluster_scMethrix}
\title{Generates a cluster object for an \code{\link{scMethrix}} object}
\usage{
cluster_scMethrix(
  scm = NULL,
  dist = NULL,
  assay = "score",
  type = c("hierarchical", "partition", "model"),
  colname = "Cluster",
  n_clusters = 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{scm}{scMethrix; Input \code{\link{scMethrix}} object. If this is specified the distance matrix will be a generic \code{\link[bioDist]{spearman.dist}} distance}

\item{dist}{dist; Optional. A distance matrix generated for an assay. Will use default paramaters for \code{\link{get_distance_matrix}}.}

\item{assay}{string; The assay to use. Default is 'score'}

\item{type}{string; The type of distance metric to use. Available options are 'hierarchical', 'partition', "model". An arbitrary cluster function can be used, and must return a named vector containing integers representing the cluster membership (e.g. \code{c(C1=1,C2=1,C3=1,C4=2)}).}

\item{colname}{string; the name of the colData column that contains the cluster information}

\item{n_clusters}{integer; the desired number of clusters. This is ignored for model-based clustering}

\item{verbose}{boolean; flag to output messages or not}

\item{...}{Additional parameters for the clustering functions}
}
\value{
An \code{\link{scMethrix}} object
}
\description{
Generates a cluster object for an \code{\link{scMethrix}} object
}
\details{
Enables multiple methods of clustering to classify samples in an \code{\link{scMethrix}} object. Either an \code{\link{scMethrix}} object or a \code{\link[stats]{dist}} object must be provided for clustering.
}
\examples{
data('scMethrix_data')
scMethrix_data <- impute_regions(scMethrix_data)
dist <- get_distance_matrix(scMethrix_data,assay = "impute")  

# For a generic clustering function
# The function must return a named vector of integers

}
\seealso{
\code{\link[=get_distance_matrix]{get_distance_matrix()}} for distance metrics, \code{\link[=hclust]{hclust()}} for heirarchical clustering, \code{\link[=kmeans]{kmeans()}} for partition clustering, \code{\link[=Mclust]{Mclust()}} for model clustering
}
