% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_transforms.R
\name{collapse_samples}
\alias{collapse_samples}
\title{Collapses multiple samples into a single sample}
\usage{
collapse_samples(
  scm = NULL,
  colname = NULL,
  trans = NULL,
  h5_dir = NULL,
  batch_size = 20,
  n_threads = 1,
  replace = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{colname}{string; The colname from \code{colData(scm)} indicating which samples should be collapse together}

\item{trans}{named vector of closures; The transforms for each assay in a named vector. Default NULL, meaning that
operations for "counts" assay is sum(x, na.rm=TRUE), and for all other assays is mean(x, na.rm=TRUE)
\code{\link{IRanges}} package.}

\item{h5_dir}{string; The directory to use. Will be created if it does not exist. Default = NULL}

\item{batch_size}{integer; The maximum number of elements to process at once.}

\item{n_threads}{integer; Maximum number of parallel instances. Default = 1}

\item{replace}{boolean; flag for whether to delete the contents of h5_dir before saving}

\item{verbose}{boolean; Flag for outputting function status messages. Default = TRUE}
}
\value{
An \code{\link{scMethrix}} object
}
\description{
Collapses multiple samples into a single sample
}
\details{
Uses the inputted function to transform an assay in the \code{\link{scMethrix}} object. Typically, most assays will use either mean (for measurements) or sum (for counts). The transform is applied column-wise to optimize how HDF5 files access sample data. If HDF5 objects are used, transform functions should be  from \pkg{DelayedMatrixStats}.

In the output object, the number of CpGs in each region is saved in mcol(scm)$n_cpgs.
}
\examples{
data('scMethrix_data')
regions <- GRanges(seqnames = c("chr1"), ranges = IRanges(1,200000000)) 
regions <- unlist(tile(regions,10))
bin_scMethrix(scMethrix_data, regions = regions)
}
