% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_operations.R
\name{convert_HDF5_scMethrix}
\alias{convert_HDF5_scMethrix}
\title{Converts HDF5 \code{\link{scMethrix}} object to an in-memory \code{\link{scMethrix}} object.}
\usage{
convert_HDF5_scMethrix(scm = NULL, verbose = TRUE)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{verbose}{boolean; Flag for outputting function status messages. Default = TRUE}
}
\value{
An object of class \code{\link{scMethrix}}
}
\description{
Converts HDF5 \code{\link{scMethrix}} object to an in-memory \code{\link{scMethrix}} object.
}
\details{
Takes an HDF\%-based \code{\link{scMethrix}} object and returns with the same object with in-memory assay slots.
}
\examples{
data('scMethrix_data')
dir <- paste0(tempdir(),"/h5")
scm <- convert_scMethrix(scMethrix_data)
convert_HDF5_scMethrix(scm)
}
