% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_operations.R
\name{convert_scMethrix}
\alias{convert_scMethrix}
\title{Converts an in-memory \code{\link{scMethrix}} to an HDF5 \code{\link{scMethrix}}}
\usage{
convert_scMethrix(
  scm = NULL,
  type = c(NA, "HDF5", "memory"),
  h5_dir = NULL,
  verbose = TRUE
)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{type}{string; what type of scMethrix to convert to. If NULL, this will convert to the opposite type, otherwise, will convert (if necessary) to the type specified}

\item{h5_dir}{string; The directory to store HDF5 files. Will be created if it does not exist. Default = NULL}

\item{verbose}{boolean; Flag for outputting function status messages. Default = TRUE}
}
\value{
An object of class \code{\link{scMethrix}}, HDF5 format
}
\description{
Converts an in-memory \code{\link{scMethrix}} to an HDF5 \code{\link{scMethrix}}
}
\details{
Takes a \code{\link{scMethrix}} object and returns with the same object with delayed array assay slots
with HDF5 backend. Might take long time!
}
\examples{
data('scMethrix_data')
convert_scMethrix(scMethrix_data, h5_dir=paste0(tempdir(),"/h5"))
}
