% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_operations.R
\name{convert_to_methrix}
\alias{convert_to_methrix}
\title{Converts an \code{\link{scMethrix}} object to methrix object}
\usage{
convert_to_methrix(scm = NULL, h5_dir = NULL)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{h5_dir}{Location to save the methrix H5 file}
}
\value{
a \code{\link[methrix]{methrix}} object
}
\description{
Converts an \code{\link{scMethrix}} object to methrix object
}
\details{
Removes extra slot data from an \code{\link{scMethrix}} object and changes structure to match
\code{\link[methrix]{methrix}} format
}
\examples{
data('scMethrix_data')
# convert_to_methrix(scMethrix_data)
}
