% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_operations.R
\name{coverage_filter}
\alias{coverage_filter}
\title{Filter matrices by coverage}
\usage{
coverage_filter(
  scm = NULL,
  cov_thr = 1,
  min_samples = NULL,
  prop_samples = NULL,
  group = NULL,
  n_chunks = 1,
  n_threads = 1
)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{cov_thr}{minimum coverage required to call a loci covered}

\item{min_samples}{Minimum number of samples that should have a loci with coverage >= \code{cov_thr}. If \code{group} is given, then this applies per group. Only need one of \code{prop_samples} or \code{min_samples}.}

\item{prop_samples}{Minimum proportion of samples that should have a loci with coverage >= \code{cov_thr}. If \code{group} is given, then this applies per group. Only need one of \code{prop_samples} or \code{min_samples}.}

\item{group}{a column name from sample annotation that defines groups. In this case, the number of min_samples will be
tested group-wise.}

\item{n_chunks}{integer; Number of chunks to split the \code{\link{scMethrix}} object in case it is very large. Default = 1}

\item{n_threads}{integer; Number of parallel instances. \code{n_threads} must be less than or equal to \code{n_chunks}. Default = 1}
}
\value{
An object of class \code{\link{scMethrix}}
}
\description{
Filter matrices by coverage
}
\details{
Takes \code{\link{scMethrix}} object and filters CpGs based on coverage statistics
}
\examples{
data('scMethrix_data')
coverage_filter(scMethrix_data, min_sample=2)
}
