% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_object.R
\name{create_scMethrix}
\alias{create_scMethrix}
\title{Create an scMethrix object}
\usage{
create_scMethrix(
  assays = NULL,
  colData = NULL,
  rowRanges = NULL,
  is_hdf5 = FALSE,
  genome_name = "hg19",
  chrom_size = NULL,
  desc = NULL,
  h5_dir = NULL,
  replace = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{assays}{list of matrices; The assays to include in the experiment}

\item{colData}{data.frame; the data corresponding to each sample in the assays}

\item{rowRanges}{GenomicRanges; the genomic loci corresponding to the assays}

\item{is_hdf5}{boolean; flag for HDF5 format}

\item{genome_name}{string; the name of the genome build used}

\item{chrom_size}{named list of integers; the size of each chromosome in the genomic ranges}

\item{desc}{named list of strings; list of relevent experiment data}

\item{h5_dir}{string; The directory to store HDF5 files. Will be created if it does not exist. Default = NULL}

\item{replace}{boolean; flag for whether to delete the contents of \code{h5_dir} before saving}

\item{verbose}{boolean; Flag for outputting function status messages. Default = TRUE}
}
\description{
Create an scMethrix object
}
