% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_funcs.R
\name{.validateAssay}
\alias{.validateAssay}
\title{Validates an assay is in the object. Allows partial pattern.}
\usage{
.validateAssay(scm = NULL, assay = NULL, throw = TRUE)
}
\arguments{
\item{scm}{scMethrix; the experiment object}

\item{assay}{string; the name of the assay}

\item{throw}{boolean; throw an error if true, return false if not}
}
\value{
string; the name of the matched assay
}
\description{
Validates an assay is in the object. Allows partial pattern.
}
\details{
Check the assays in an scMethrix object and partial matches
}
