% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs.R
\name{.validateThreads}
\alias{.validateThreads}
\title{Validates the number of threads for the session. Windows can only support one thread}
\usage{
.validateThreads(n_threads)
}
\arguments{
\item{n_threads}{numeric; the number of threads}
}
\value{
integer; 1 if windows, or some number of threads between 1 and parallel::detectCores
}
\description{
Validates the number of threads for the session. Windows can only support one thread
}
