% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs.R
\name{.validateValue}
\alias{.validateValue}
\title{Validates numeric values based on some experession}
\usage{
.validateValue(value, ...)
}
\arguments{
\item{value}{numeric; the value to test}

\item{...}{string; the expressions to test}
}
\value{
invisible(TRUE), if the object is valid. Error if not.
}
\description{
Validates numeric values based on some experession
}
