% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_operations.R
\name{export_bed}
\alias{export_bed}
\title{Exports all samples in an \code{\link{scMethrix}} objects into individual bedgraph files}
\usage{
export_bed(scm = NULL, path = NULL, suffix = NULL, verbose = TRUE)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{path}{the \code{\link{file.path}} of the directory to save the files}

\item{suffix}{optional suffix to add to the exported bed files}

\item{verbose}{boolean; Flag for outputting function status messages. Default = TRUE}
}
\value{
nothing
}
\description{
Exports all samples in an \code{\link{scMethrix}} objects into individual bedgraph files
}
\examples{
data('scMethrix_data')
export_bed(scMethrix_data,path=paste0(tempdir(),"/export"))
}
