% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_export.R
\name{export_beds}
\alias{export_beds}
\title{Exports all samples in an \code{\link{scMethrix}} objects into individual bedgraph files}
\usage{
export_beds(
  scm = NULL,
  path = NULL,
  suffix = NULL,
  verbose = TRUE,
  include = FALSE,
  na.rm = TRUE,
  header = FALSE
)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{path}{character; the \code{\link{file.path}} of the directory to save the files}

\item{suffix}{character; optional suffix to add to the exported bed files}

\item{verbose}{boolean; Flag for outputting function status messages. Default = TRUE}

\item{include}{boolean; flag to include the values of non-standard assays in the bedgraph file}

\item{na.rm}{boolean; flag to remove the NA values from the output data}

\item{header}{boolean; flag to add the header onto each column}
}
\value{
nothing
}
\description{
Exports all samples in an \code{\link{scMethrix}} objects into individual bedgraph files
}
\details{
The structure of the bedgraph files will be a tab-deliminated structure of:
Chromosome | CpG start site | CpG end site | methylation score | coverage | Additional assays (if include = TRUE)

If additional assays are used, and headers enabled, it is up to the user to ensure that assay names are not protected in any downstream analysis of the bedgraph files
}
\examples{
data('scMethrix_data')
export_beds(scMethrix_data,path=paste0(tempdir(),"/export"))
}
