% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_export.R
\name{export_bigwigs}
\alias{export_bigwigs}
\title{Exports scMethrix object as bigWigs}
\usage{
export_bigwigs(scm, assay = "score", path = tempdir(), samp_names = NULL)
}
\arguments{
\item{scm}{\code{\link{scMethrix}} object}

\item{assay}{string; the assay to export. Default is "score"}

\item{path}{string; Output directory name where the files should be saved. Default tempdir()}

\item{samp_names}{string; List of sample names to export}
}
\description{
Exports scMethrix object as bigWigs
}
\examples{
\dontrun{
data('scMethrix_data')
export_bigwigs(scm = scMethrix_data, assay = "score", output_dir = tempdir())
}
}
