% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_export.R
\name{export_bsseq}
\alias{export_bsseq}
\title{Convert \code{\link{scMethrix}} to \code{bsseq} object}
\usage{
export_bsseq(scm, m_assay = "score", c_assay = "counts")
}
\arguments{
\item{scm}{\code{\link{methrix}} object}

\item{m_assay}{matrix; the assay containing methylation scores}

\item{c_assay}{matrix; the assay containing count scores}
}
\value{
An object of class \code{bsseq}
}
\description{
Convert \code{\link{scMethrix}} to \code{bsseq} object
}
\details{
Takes \code{\link{scMethrix}} object and returns a \code{bsseq} object.
}
\examples{
\dontrun{
data('scMethrix_data')
export_bsseq(scMethrix_data)
}
}
