% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_export.R
\name{export_methrix}
\alias{export_methrix}
\title{Converts an \code{\link{scMethrix}} object to methrix object}
\usage{
export_methrix(scm = NULL, h5_dir = NULL)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{h5_dir}{Location to save the methrix H5 file}
}
\value{
a \code{\link[methrix]{methrix}} object
}
\description{
Converts an \code{\link{scMethrix}} object to methrix object
}
\details{
Removes extra slot data from an \code{\link{scMethrix}} object and changes structure to match
\code{\link[methrix]{methrix}} format. A 'counts' assay for coverage values must be present.
Functionality not supported by methrix (e.g. reduced dimensionality)
will be discarded.
}
\examples{
data('scMethrix_data')
# convert_to_methrix(scMethrix_data)
}
