% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_funcs.R
\name{fill}
\alias{fill}
\title{Fills a vector with a specified \code{fill} value}
\usage{
fill(x, val = 0)
}
\arguments{
\item{x}{vector; A vector in which to fill the NA values}

\item{val}{basic data type; Any value from one of R's basic data types (character, numeric, integer, logical, complex)}
}
\value{
vector; Same values as input vector, but NA values are replaced with \code{fill} if above of below the threshold, or 'rep.na' if NA
}
\description{
Fills a vector with a specified \code{fill} value
}
\examples{
vals <- c(0,0.25,0.5,0.75,1,NA)
fill(vals, val=2)
}
