% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_operations.R
\name{filter_scMethrix}
\alias{filter_scMethrix}
\title{Filters an \code{\link{scMethrix}} object based on given conditions.}
\usage{
filter_scMethrix(
  m,
  regions = NULL,
  contigs = NULL,
  samples = NULL,
  by = c("include", "exclude"),
  verbose = TRUE
)
}
\arguments{
\item{m}{\code{\link{scMethrix}} object}

\item{regions}{genomic regions to subset by. Could be a data.table with 3 columns (chr, start, end) or a \code{GenomicRanges} object}

\item{contigs}{string of chromosome names to subset by}

\item{samples}{string of sample names to subset by}

\item{verbose}{flag to output messages or not}
}
\value{
An object of class \code{\link{scMethrix}}
}
\description{
Filters an \code{\link{scMethrix}} object based on given conditions.
}
\details{
Takes \code{\link{scMethrix}} object and filters CpGs based on region, contig and/or sample. Can
either subset to or filter out the input parameters.
}
\examples{
data('scMethrix_data')

contigs <- c("chr1","chr3")
regions <- GenomicRanges::GRanges(seqnames = "chr1", ranges = IRanges(1,5)) 
samples <- c("df1","df3")

#Subset to only chromosome 1
filter_scMethrix(scMethrix_data$mem, contigs = contigs)

#Subset to only samples bed1 and bed3
filter_scMethrix(scMethrix_data$mem, samples = samples)

#Subset to only samples bed1 and bed3, and chromosome 1
filter_scMethrix(scMethrix_data$mem, samples = samples, contigs = contigs)

#Subset to only region "chr1:1-5"
filter_scMethrix(scMethrix_data$mem, regions = regions)

#Filter to exclude chromosome 1
filter_scMethrix(scMethrix_data$mem, contigs = contigs, by = "exclude")

#Filter to exclude samples bed1 and bed3
filter_scMethrix(scMethrix_data$mem, samples = samples, by = "exclude")

#Filter to exclude samples bed1 and bed3, and chromosome 1
filter_scMethrix(scMethrix_data$mem, samples = samples, contigs = contigs, by = "exclude")

#Filter to exclude region "chr1:1-5"
filter_scMethrix(scMethrix_data$mem, regions = regions, by = "exclude")
}
