% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_transforms.R
\name{generate_random_subset}
\alias{generate_random_subset}
\title{Generates a random subset of CpG sites}
\usage{
generate_random_subset(scm = NULL, n_cpgs = 10000, seed = "123")
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{n_cpgs}{numeric; The number of CpGs to include}

\item{seed}{string; value to use for sampling}
}
\value{
scMethrix; an experiment with n_cpgs
}
\description{
Generates a random subset of CpG sites
}
\details{
From an \code{\link{scMethrix}} object, this will randomly select \code{n_cpgs} and create a new
object containing only those CpGs. This is typically used for approximation or visualization. The seed
can be specified for consistency.
}
\examples{
data('scMethrix_data')
generate_random_subset(scMethrix_data,n_cpgs = round(nrow(scMethrix_data)/2))
}
