% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_object.R
\name{generic_scMethrix_function}
\alias{generic_scMethrix_function}
\title{Function used only for inheritence for Roxygen2 documentation. Lists the common function inputs used in the package}
\usage{
generic_scMethrix_function(
  scm,
  assay,
  new_assay,
  trans,
  verbose,
  n_chunks,
  n_threads,
  h5_dir
)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{assay}{string; name of an existing assay. Default = "score"}

\item{new_assay}{string; name for transformed assay. Default = "new_assay"}

\item{trans}{closure; The transformation function. Default = mean}

\item{verbose}{boolean; Flag for outputting function status messages. Default = TRUE}

\item{n_chunks}{integer; Number of chunks to split the \code{\link{scMethrix}} object in case it is very large. Default = 1}

\item{n_threads}{integer; Number of parallel instances. \code{n_threads} must be less than or equal to \code{n_chunks}. Default = 1}

\item{h5_dir}{string; The directory to use. Will be created if it does not exist. Default = NULL}
}
\description{
Function used only for inheritence for Roxygen2 documentation. Lists the common function inputs used in the package
}
