% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_operations.R
\name{get_coldata_stats}
\alias{get_coldata_stats}
\title{Adds descriptive statistics to colData columns in an \code{\link{scMethrix}} object.}
\usage{
get_coldata_stats(scm, assay = "score", suffix = "")
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{assay}{string; name of an existing assay. Default = "score"}

\item{suffix}{string; a suffix to add to the string}
}
\value{
An \code{\link{scMethrix}} object
}
\description{
Adds descriptive statistics to colData columns in an \code{\link{scMethrix}} object.
}
\details{
Adds the mean, SD, and sample count for each sample in an \code{\link{scMethrix}} object. This can be accessed using colData(). Columns with the names of 'mean','sd', and 'cpg' will be automatically overwritten, but \code{suffix} can be used to keep multiple stats columns.

This data will not be updated automatically for any subset, merge, bin, etc functions.
}
\examples{
data('scMethrix_data')
get_coldata_stats(scMethrix_data)
}
