% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_clustering.R
\name{get_distance_matrix}
\alias{get_distance_matrix}
\title{Get the pair-wise distance matrix for an assay}
\usage{
get_distance_matrix(
  scm,
  assay = "score",
  type = c("pearson", "spearman", "kendall", "euclidean", "manhattan", "canberra",
    "binary", "minkowski"),
  verbose = TRUE
)
}
\arguments{
\item{scm}{scMethrix; Input \code{\link{scMethrix}} object}

\item{assay}{string; The assay to use. Default is 'score'}

\item{type}{string; The type of distance metric to use. Available options are "pearson", "spearman", "tau", "euclidean", "manhattan", "canberra", "binary", "minkowski". An aribitrary distance function can also be used, so long as the input takes just the specified matrix.}

\item{verbose}{boolean; flag to output messages or not}
}
\value{
matrix; the distance matrix
}
\description{
Get the pair-wise distance matrix for an assay
}
\details{
Utilizes mainly the bioDist package to determine various distance metrics to be used for later clustering.
}
\examples{
data('scMethrix_data')
scMethrix_data <- impute_regions(scMethrix_data) 
get_distance_matrix(scMethrix_data,assay = "impute") 

# For an arbitrary distance function
library(bioDist)
fun <- bioDist::spearman.dist
get_distance_matrix(scMethrix_data,assay = "impute",type = bioDist::spearman.dist)
}
\seealso{
\url{https://www.rdocumentation.org/packages/stats/versions/3.6.2/topics/dist}

\url{https://www.bioconductor.org/packages//2.7/bioc/html/bioDist.html}
}
