% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_operations.R
\name{get_matrix}
\alias{get_matrix}
\title{Extract methylation or coverage matrices}
\usage{
get_matrix(m, add_loci = FALSE, in_granges = FALSE)
}
\arguments{
\item{m}{\code{\link{scMethrix}} object}

\item{add_loci}{Default FALSE. If TRUE adds CpG position info to the matrix and returns as a data.table}

\item{in_granges}{Do you want the outcome in \code{GRanges}?}

\item{type}{can be \code{M} or \code{C}. Default 'M'}
}
\value{
Coverage or Methylation matrix
}
\description{
Extract methylation or coverage matrices
}
\details{
Takes \code{\link{scMethrix}} object and returns user specified \code{methylation} or \code{coverage} matrix
}
