% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_operations.R
\name{get_metadata_stats}
\alias{get_metadata_stats}
\title{Add descriptive statistics to metadata columns in an \code{\link{scMethrix}} object.}
\usage{
get_metadata_stats(m)
}
\arguments{
\item{m}{A \code{\link{scMethrix}} object}
}
\value{
An \code{\link{scMethrix}} object
}
\description{
Add descriptive statistics to metadata columns in an \code{\link{scMethrix}} object.
}
\examples{
data('scMethrix_data')
get_metadata_stats(scMethrix_data)
}
