% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_plot.R
\name{get_palette}
\alias{get_palette}
\title{Getter for plot palette colors}
\usage{
get_palette(n_row, col_palette)
}
\arguments{
\item{n_row}{Number of colors}

\item{col_palette}{String for RColorBrewer palette name}
}
\value{
RColorBrewer palette
}
\description{
Getter for plot palette colors
}
