% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_operations.R
\name{get_region_summary}
\alias{get_region_summary}
\title{Extracts and summarizes methylation or coverage info by regions of interest}
\usage{
get_region_summary(
  scm = NULL,
  assay = "score",
  regions = NULL,
  group = NULL,
  n_chunks = 1,
  n_threads = 1,
  by = c("mean", "median", "maximum", "minimum", "sum", "sd"),
  overlap_type = c("within", "start", "end", "any", "equal"),
  verbose = TRUE
)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{assay}{string; name of an existing assay. Default = "score"}

\item{regions}{GRanges;  genomic regions to be summarized. Could be a data.table with 3 columns (chr, start, end) or a \code{\link{GenomicRanges}} object}

\item{group}{a column name from sample annotation that defines groups. In this case, the number of min_samples will be tested group-wise.}

\item{n_chunks}{integer; Number of chunks to split the \code{\link{scMethrix}} object in case it is very large. Default = 1}

\item{n_threads}{integer; Maximum number of parallel instances. Default = 1}

\item{by}{closure; mathematical function by which regions should be summarized. Can be one of the following: mean, sum, max, min. Default 'mean'}

\item{overlap_type}{defines the type of the overlap of the CpG sites with the target region. Default value is \code{within}. For detailed description, see the \code{findOverlaps} function of the \code{\link{IRanges}} package.}

\item{verbose}{boolean; Flag for outputting function status messages. Default = TRUE}
}
\value{
table of summary statistic for the given region
}
\description{
Extracts and summarizes methylation or coverage info by regions of interest
}
\details{
Summarizes regions and/or groups for descriptive statistics.
}
\examples{
data('scMethrix_data')

# Determine global methylation of groups
# summarize_stats(scMethrix_data, assay = 'score', group = "Group",  by = 'mean') 
#TODO: update the scMethrix_data for group info

# Determine methylation status of chromosomes
#summarize_stats(scMethrix_data, assay = "score", regions = range(rowRanges(scMethrix_data)))

# Determine methylation status of chromosomes by groups
#summarize_stats(scMethrix_data, assay = "score", group = "Group", 
#regions = range(rowRanges(scMethrix_data)))
}
