% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_operations.R
\name{get_rowdata_stats}
\alias{get_rowdata_stats}
\title{Adds descriptive statistics to metadata columns in an \code{\link{scMethrix}} object.}
\usage{
get_rowdata_stats(scm, assay = "score", suffix = "")
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{assay}{string; name of an existing assay. Default = "score"}

\item{suffix}{string; a suffix to add to the string}
}
\value{
An \code{\link{scMethrix}} object
}
\description{
Adds descriptive statistics to metadata columns in an \code{\link{scMethrix}} object.
}
\details{
Adds the mean, median, SD, and sample count and coverage (if present) for  the \code{\link{GenomicRanges}} in an \code{\link{scMethrix}} object. This can be accessed using mcols().

This data will not be updated automatically for any subset, merge, bin, etc functions.
}
\examples{
data('scMethrix_data')
get_rowdata_stats(scMethrix_data)
}
