% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_plot.R
\name{get_shape}
\alias{get_shape}
\title{Getter for plot shapes. Shapes selected for optimal distinction and taken from:
http://www.sthda.com/english/wiki/r-plot-pch-symbols-the-different-point-shapes-available-in-r}
\usage{
get_shape(n_row)
}
\arguments{
\item{n_row}{Number of shapes. Max of 15.}
}
\value{
list of shapes (by integer)
}
\description{
Getter for plot shapes. Shapes selected for optimal distinction and taken from:
http://www.sthda.com/english/wiki/r-plot-pch-symbols-the-different-point-shapes-available-in-r
}
