% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_funcs.R
\name{has_cov}
\alias{has_cov}
\title{Checks if \code{\link{scMethrix}} object has a coverage matrix}
\usage{
has_cov(scm)
}
\arguments{
\item{m}{scMethrix; The \code{\link{scMethrix}} object}
}
\value{
boolean Whether the object has a coverage matrix
}
\description{
Checks if \code{\link{scMethrix}} object has a coverage matrix
}
\examples{
data('scMethrix_data')
has_cov(scMethrix_data)
}
