% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_funcs.R
\name{?}
\alias{?}
\title{Ternary operator}
\usage{
`?`(x, y)
}
\arguments{
\item{x}{The value to return if true}

\item{y}{The value to return if false}
}
\value{
The x or y value depending on boolean
}
\description{
Ternary operator
}
\details{
Acts in the same way as ternary operators in other language. Must surround with brackets to use)
}
