% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_transforms.R
\name{impute_by_melissa}
\alias{impute_by_melissa}
\title{Imputes the NA values of a \code{\link{scMethrix}} object.}
\usage{
impute_by_melissa(scm, threshold = 50)
}
\arguments{
\item{scm}{A \code{\link{scMethrix}} object}

\item{threshold}{The value for cutoff in the "score" assay to determine methylated or unmethylated status.
Default = 50}
}
\value{
An \code{\link{scMethrix}} object
}
\description{
Imputes the NA values of a \code{\link{scMethrix}} object.
}
\details{
Uses the inputted function to transform an assay in the \code{\link{scMethrix}} object
}
\examples{
data('scMethrix_data')
}
