% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_transforms.R
\name{impute_by_RF}
\alias{impute_by_RF}
\title{Imputes the an \code{\link{scMethrix}}object using a random forest model}
\usage{
impute_by_RF(scm = NULL, assay = "score", new_assay = "impute", ...)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{assay}{string; name of an existing assay. Default = "score"}

\item{new_assay}{string; name for transformed assay. Default = "new_assay"}

\item{...}{Additional arguments for \link[missForest]{missForest}}
}
\value{
An \code{\link{scMethrix}} object
}
\description{
Imputes the an \code{\link{scMethrix}}object using a random forest model
}
\details{
Uses the inputted function to transform an assay in the \code{\link{scMethrix}} object.
}
\examples{
data('scMethrix_data')
impute_by_RF(scMethrix_data, assay = "score", new_assay = "impute")
}
\references{
Stekhoven, D. J., & Bühlmann, P. (2012). MissForest—non-parametric missing value imputation for mixed-type data. Bioinformatics, 28(1), 112-118.
}
