% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_transforms.R
\name{impute_regions}
\alias{impute_regions}
\title{Generic imputation return function}
\usage{
impute_regions(
  scm = NULL,
  assay = "score",
  new_assay = NULL,
  regions = NULL,
  op = NULL,
  n_chunks = 1,
  n_threads = 1,
  overlap_type = "within"
)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{assay}{string; name of an existing assay. Default = "score"}

\item{new_assay}{string; name for transformed assay. Default = "new_assay"}

\item{regions}{Granges; the regions to impute. Default is by chromosome.}

\item{op}{closure; the imputation operation}

\item{n_chunks}{integer; Number of chunks to split the \code{\link{scMethrix}} object in case it is very large. Default = 1}

\item{n_threads}{integer; Number of parallel instances. \code{n_threads} should be less than or equal to \code{n_chunks}. Default = 1}

\item{overlap_type}{string;}
}
\value{
list; two \code{\link{scMethrix}} objects names 'training' and 'test'
}
\description{
Generic imputation return function
}
\details{
Uses the specified imputation operation to evaluation an scMethrix object.
}
\examples{
data('scMethrix_data')
generate_training_set(scMethrix_data, training_prop = 0.2)
}
