% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_funcs.R
\name{is_h5}
\alias{is_h5}
\title{Checks if scMethrix object is an HDF5 object}
\usage{
is_h5(m)
}
\arguments{
\item{m}{The scMethrix object}
}
\value{
boolean Whether the object is HDF5
}
\description{
Checks if scMethrix object is an HDF5 object
}
\details{
Acts in the same way as ternary operators in other language. Must surround with brackets to use)
}
