% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_ops.R
\name{liftover_CpGs}
\alias{liftover_CpGs}
\title{Converts from one reference to another via rtracklayer::liftOver}
\usage{
liftover_CpGs(scm, chain = NULL, target_genome = NULL, verbose = TRUE)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{chain}{string; the file location for the desired chain to use in liftOver}

\item{target_genome}{string; the target genome. This will be update the genome field in the output scMethrix object}

\item{verbose}{boolean; Flag for outputting function status messages. Default = TRUE}
}
\value{
a list of data.table containing number of CpG's and contig lengths
}
\description{
Converts from one reference to another via rtracklayer::liftOver
}
\details{
This conversion is one-to-many, so for consistency, only the first element in the target assembly is used.

LiftOver chains can be found here: \href{https://hgdownload.soe.ucsc.edu/downloads.html}{UCSC}

Here's links to common chains:
{http://hgdownload.cse.ucsc.edu/goldenpath/hg38/liftOver/hg38ToHg19.over.chain.gz}{hg38 to hg19}
{https://hgdownload.soe.ucsc.edu/gbdb/hg19/liftOver/hg19ToHg38.over.chain.gz}{hg19 to hg38}

AnnotationHub can also be used to download chains. See an example in the vignettes:
}
\examples{
\dontrun{
hg19_cpgs = methrix::extract_CPGs(ref_genome = 'BSgenome.Hsapiens.UCSC.hg19')
}
}
