% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_operations.R
\name{load_HDF5_scMethrix}
\alias{load_HDF5_scMethrix}
\title{Loads HDF5 scMethrix object}
\usage{
load_HDF5_scMethrix(dir = NULL, ...)
}
\arguments{
\item{dir}{The directory to read in from. Default NULL}

\item{...}{Parameters to pass to \code{\link{loadHDF5SummarizedExperiment}}}
}
\value{
An object of class \code{\link{scMethrix}}
}
\description{
Loads HDF5 scMethrix object
}
\details{
Takes  directory with a previously saved HDF5Array format \code{\link{scMethrix}} object and loads it
}
\examples{
data('scMethrix_data')
dir <- paste0(tempdir(),"/h5")
scm <- convert_scMethrix(scMethrix_data, h5_dir=dir)
save_HDF5_scMethrix(scm, h5_dir = dir, replace = TRUE)
n <- load_HDF5_scMethrix(dir)
}
