% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_operations.R
\name{load_scMethrix}
\alias{load_scMethrix}
\title{Loads HDF5 \code{\link{scMethrix}} object}
\usage{
load_scMethrix(dest = NULL, verbose = TRUE, ...)
}
\arguments{
\item{dest}{The directory or file to read in from}

\item{verbose}{boolean; Flag for outputting function status messages. Default = TRUE}

\item{...}{Parameters to pass to \code{\link{loadHDF5SummarizedExperiment}}}
}
\value{
An object of class \code{\link{scMethrix}}
}
\description{
Loads HDF5 \code{\link{scMethrix}} object
}
\details{
Takes directory with a previously saved HDF5Array format \code{\link{scMethrix}} object and loads it
}
\examples{
data('scMethrix_data')
dir <- paste0(tempdir(),"/h5")
scm <- convert_scMethrix(scMethrix_data, h5_dir=dir)
save_scMethrix(scm, dest = dir, replace = TRUE)
n <- load_scMethrix(dest = dir)
}
