% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_operations.R
\name{mask_by_coverage}
\alias{mask_by_coverage}
\title{Masks CpGs by coverage}
\usage{
mask_by_coverage(
  scm = NULL,
  assay = "score",
  low_threshold = 0,
  avg_threshold = 2,
  n_threads = 1,
  verbose = TRUE
)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{assay}{string; name of an existing assay. Default = "score"}

\item{low_threshold}{numeric; The minimal coverage allowed. Everything below will get masked. If NULL, this will be ignored. Default = 0}

\item{avg_threshold}{numeric; The max average coverage. If NULL, this will be ignored. Default = 2}

\item{n_threads}{integer; Number of parallel instances. Can only be used if \code{\link{scMethrix}} is in HDF5 format. Default = 1}

\item{verbose}{boolean; Flag for outputting function status messages. Default = TRUE}
}
\value{
An object of class \code{\link{scMethrix}}
}
\description{
Masks CpGs by coverage
}
\details{
Takes \code{\link{scMethrix}} object and masks sites with low overall or high average coverage by putting NA for assay values. The sites will remain in the object and all assays will be affected.

\code{low_threshold} is used to mask sites with low overall coverage.
\code{avg_threshold} is used to mask sites with high aberrant counts. For single cell data, this is typically CpG sites with an average count > 2, as there are only two strands in a cell to sequence.
}
\examples{
data('scMethrix_data')
mask_by_coverage(scMethrix_data,low_threshold=2, avg_threshold=2)
}
