% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_operations.R
\name{mask_by_idx}
\alias{mask_by_idx}
\title{Specified rows and columns are masked with NA values}
\usage{
mask_by_idx(scm, col_idx = NULL, row_idx = NULL, verbose = TRUE)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{col_idx}{numeric; A vector of column indexes to replace all values with NA}

\item{row_idx}{numeric; A vector of row indexes to replace all values with NA}

\item{verbose}{boolean; Flag for outputting function status messages. Default = TRUE}
}
\value{
An object of class \code{\link{scMethrix}}
}
\description{
Specified rows and columns are masked with NA values
}
\details{
This iterates through all assays in the inputted \code{\link{scMethrix}} object and replaces all rows in row_idx and cols in col_idx.
}
\seealso{
\code{\link[=mask_scMethrix]{mask_scMethrix()}} wraps this function and generates idxs by statistics, \code{\link[=remove_uncovered]{remove_uncovered()}} to remove the masked sites

Other quality control: 
\code{\link{mask_scMethrix}()}
}
\concept{masking}
\concept{quality control}
