% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_operations.R
\name{mask_scMethrix}
\alias{mask_scMethrix}
\title{Masks rows or columns based on some descriptive statistic}
\usage{
mask_scMethrix(
  scm = NULL,
  assay = "score",
  threshold = 0,
  by = c("row", "column"),
  stat = c("count", "sum", "mean", "median", "sd", "variance", "proportion"),
  op = c("==", "!=", ">", "<", ">=", "<="),
  na.rm = TRUE,
  n_threads = 1,
  verbose = TRUE
)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{assay}{string; name of an existing assay. Default = "score"}

\item{threshold}{numeric; The threshold value to compute from}

\item{by}{string; Calculate over rows or columns}

\item{stat}{string; The calculation to perform on each row}

\item{op}{string; The operator to compare the calculation to the threshold}

\item{na.rm}{boolean; flag to remove NA values}

\item{n_threads}{integer; Maximum number of parallel instances. Default = 1}

\item{verbose}{boolean; Flag for outputting function status messages. Default = TRUE}
}
\value{
An object of class \code{\link{scMethrix}}
}
\description{
Masks rows or columns based on some descriptive statistic
}
\details{
Takes \code{\link{scMethrix}} object and masks CpG sites based on row statistics. The sites will remain in the object and all assays will be masked. These sites can later be removed with \code{\link[=remove_uncovered]{remove_uncovered()}}.
\subsection{Types of functions}{

This is a very flexible function, and can do operations like:
\itemize{
\item Remove samples with few CpG sites
\item Remove CpGs with low coverage or are not present in many cells
\item Remove low variance CpG sites
}
}

\subsection{Notes}{
\itemize{
\item For \code{stat = "variance"}, a CpG that is either hypo- or hyper-methylated in all samples will have a variability of 0, whereas a CpG that is exactly half of each will have a value of 1.
\item For \code{stat = "variance"} and \code{stat = "sd"}, CpGs present in only one sample will automatically have a variance/SD of 0
@family masking
}
}
}
\examples{
data('scMethrix_data')

## Mask by low CpG count
# This will mask any samples that have < 170 total CpGs present
mask_scMethrix(scMethrix_data, assay="score", threshold = 170, by = "col", stat="count", op="<")

#' ## Mask by low sample count
# This will remove CpGs that are not present in > 2 samples
mask_scMethrix(scMethrix_data, assay="score", threshold=2, by = "row", stat="count", op="<=")

## Mask by low coverage
# This will mask each CpG in the counts assay where the total coverage is < 5
mask_scMethrix(scMethrix_data, assay="counts", threshold=5, by = "row", stat="sum", op="<")

## Mask by high average coverage
# This will mask each CpG in the counts assay where the average coverage is > 1
mask_scMethrix(scMethrix_data, assay="counts", threshold=1, by = "row", stat="mean", op=">")

## Mask by variance
# This will mask low variance (homogenous) CpG sites in the score assay where var < 0.05
mask_scMethrix(scMethrix_data, assay="score", threshold=0.05, by = "row", stat="var", op="<")
}
\seealso{
\code{\link[=mask_by_idx]{mask_by_idx()}}, the actual masking function

Other quality control: 
\code{\link{mask_by_idx}()}
}
\concept{quality control}
