% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_funcs.R
\name{normalize}
\alias{normalize}
\title{Fills a vector with a specified \code{fill} value}
\usage{
normalize(x, min = NULL, max = NULL, scale = FALSE)
}
\arguments{
\item{x}{vector; A vector in which to fill the NA values}

\item{min}{numeric; the minimum value to normalize to}

\item{max}{numeric; the maximum value to normalize to}

\item{scale}{boolean; should the numbers normalize to 0-1, or scaled to min-max}
}
\value{
vector; the normalized vector
}
\description{
Fills a vector with a specified \code{fill} value
}
\examples{
vals <- c(0,1,2,3.4,5)

#Normalize to 0,1
normalize(vals)

#Normalize to 0,1, but the known min and max are 0-10 for the input data
normalize(vals, min = 0, max = 10)

#Normalize to 0-10
normalize(vals, min = 0, max = 10, scale = TRUE)
}
