% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_operations.R
\name{order_by_sd}
\alias{order_by_sd}
\title{Order scMethrix object by SD}
\usage{
order_by_sd(scm, zero.rm = FALSE, na.rm = FALSE)
}
\arguments{
\item{scm}{\code{\link{scMethrix}} object}

\item{zero.rm}{Removes zero values from equations (the default empty value for sparse matrices)}

\item{na.rm}{Removes the NA values from equations}
}
\value{
An object of class \code{\link{scMethrix}}
}
\description{
Order scMethrix object by SD
}
\details{
Takes \code{\link{scMethrix}} object and reorganizes the data by standard deviation
}
