% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_plot.R
\name{plot_coverage}
\alias{plot_coverage}
\title{Coverage QC Plots}
\usage{
plot_coverage(
  scm = NULL,
  type = c("histogram", "density"),
  pheno = NULL,
  perGroup = FALSE,
  lim = 100,
  size.lim = 1e+06,
  col_palette = "RdYlGn",
  show_legend = TRUE
)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{type}{string; Choose between 'histogram' (histogram) or 'density' (density plot).}

\item{pheno}{string; Col name of colData(m). Will be used as a factor to color different groups}

\item{perGroup}{boolean; Color the plots in a sample-wise manner?}

\item{lim}{integer; Maximum coverage value to be plotted.}

\item{size.lim}{integer; The maximum number of observarions (sites*samples) to use. If the dataset is larger that this,
random sites will be selected from the genome.}

\item{col_palette}{string; Name of the RColorBrewer palette to use for plotting.}

\item{show_legend}{boolean; Display the legend on the plot}
}
\value{
ggplot2 object
}
\description{
Coverage QC Plots
}
\examples{
data('scMethrix_data')
plot_coverage(scm = scMethrix_data)
}
