% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_plot.R
\name{plot_density}
\alias{plot_density}
\title{Density Plot of \eqn{\beta}-Values}
\usage{
plot_density(
  scm = NULL,
  regions = NULL,
  n_cpgs = 25000,
  pheno = NULL,
  col_palette = "RdYlGn",
  show_legend = TRUE
)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{regions}{Granges; genomic regions to be summarized. Could be a data.table with 3 columns (chr, start, end) or a \code{GenomicRanges} object}

\item{n_cpgs}{integer; Use these many random CpGs for plotting. Default 25000. Set it to \code{NULL} to use all - which can be memory expensive.}

\item{pheno}{string; Col name of colData(m). Will be used as a factor to color different groups}

\item{col_palette}{string; Name of the RColorBrewer palette to use for plotting.}

\item{show_legend}{boolean; Display the legend on the plot}
}
\value{
ggplot2 object
}
\description{
Density Plot of \eqn{\beta}-Values
}
\examples{
data('scMethrix_data')
plot_density(scm = scMethrix_data)
}
