% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_plot.R
\name{plot_dim_red}
\alias{plot_dim_red}
\title{Plot dimensionality reduction}
\usage{
plot_dim_red(
  scm,
  dim_red,
  col_anno = NULL,
  shape_anno = NULL,
  axis_labels = NULL,
  show_dp_labels = FALSE
)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{dim_red}{dimensionality reduction from an scMethrix object. Should be a matrix of two columns representing
the X and Y coordinates of the dim. red., with each row being a seperate sample}

\item{col_anno}{Column name of colData(m). Default NULL. Will be used as a factor to color different groups. Required \code{methrix} object}

\item{shape_anno}{Column name of colData(m). Default NULL. Will be used as a factor to shape different groups. Required \code{methrix} object}

\item{axis_labels}{A list of 'X' and 'Y' strings for labels, or NULL if no labels are desired}

\item{show_dp_labels}{Default FLASE}
}
\value{
ggplot2 object
}
\description{
Plot dimensionality reduction
}
