% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_plot.R
\name{plot_pca}
\alias{plot_pca}
\title{Plot PCA results}
\usage{
plot_pca(
  scm = NULL,
  col_anno = NULL,
  shape_anno = NULL,
  show_labels = FALSE,
  plot_vars = FALSE
)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{col_anno}{Column name of colData(m). Default NULL. Will be used as a factor to color different groups. Required \code{methrix} object}

\item{shape_anno}{Column name of colData(m). Default NULL. Will be used as a factor to shape different groups. Required \code{methrix} object}

\item{show_labels}{Show cell names on each data point. Default FLASE}

\item{plot_vars}{Plot the variance explanation too}
}
\value{
ggplot2 object
}
\description{
Plot PCA results
}
\examples{
data('scMethrix_data')
scmpc = pca_scMethrix(scMethrix_data)
plot_pca(scmpc)
}
\seealso{
\code{\link[=pca_scMethrix]{pca_scMethrix()}} for dimensionality reduction
}
