% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_plot.R
\name{plot_stats}
\alias{plot_stats}
\title{Plot descriptive statistics}
\usage{
plot_stats(
  scm,
  assay = "score",
  stat = c("mean", "median"),
  ignore_chr = NULL,
  samples = NULL,
  n_col = NULL,
  n_row = NULL
)
}
\arguments{
\item{scm}{scMethrix; \code{\link{get_stats}} will be run for the specified assay}

\item{assay}{string; Which assay to get the stats of. Default "score"}

\item{stat}{string; Can be \code{mean} or \code{median}. Default \code{mean}}

\item{ignore_chr}{string; Chromsomes to ignore. If NULL, all chromosome will be used. Default \code{NULL}}

\item{samples}{list of strings; Samples to ignore.  If NULL, all samples will be used. Default \code{NULL}}

\item{n_col}{integer; number of columns. Passed to \code{facet_wrap}}

\item{n_row}{integer; number of rows. Passed to \code{facet_wrap}}
}
\value{
ggplot2 object
}
\description{
Plot descriptive statistics
}
\details{
plot descriptive statistics results from \code{\link{get_stats}}
}
\examples{
data('scMethrix_data')
plot_stats(scMethrix_data)
}
\seealso{
\code{\link{get_stats}}
}
