% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_plot.R
\name{plot_tsne}
\alias{plot_tsne}
\title{Plot tSNE results}
\usage{
plot_tsne(scm)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the experiment containing the tSNE}
}
\value{
ggplot2 object
}
\description{
Plot tSNE results
}
\examples{
data('scMethrix_data')
}
