% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_plot.R
\name{plot_umap}
\alias{plot_umap}
\title{Plot UMAP results}
\usage{
plot_umap(scm)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the experiment containing the UMAP}
}
\value{
ggplot2 object
}
\description{
Plot UMAP results
}
\examples{
data('scMethrix_data')
}
