% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_funcs.R
\name{rd}
\alias{rd}
\title{Same as rowData(scm), but shorter syntax, and will output row names if there is no columns}
\usage{
rd(scm)
}
\arguments{
\item{scm}{an scMethrix}
}
\description{
Same as rowData(scm), but shorter syntax, and will output row names if there is no columns
}
