% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_beds.R
\name{read_bed_by_index}
\alias{read_bed_by_index}
\title{Parses BED files for methylation values using previously generated index genomic coordinates}
\usage{
read_bed_by_index(
  file,
  ref_cpgs = NULL,
  col_list = NULL,
  zero_based = FALSE,
  strand_collapse = FALSE,
  fill = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{file}{string; file.paths of BED file to parse}

\item{ref_cpgs}{data.table; list of CpG sites in the tab-delimited format of chr-start-end. Must be zero-based genome.}

\item{col_list}{string; The column index object for the input BED files}

\item{zero_based}{boolean; flag for whether the input data is zero-based or not}

\item{strand_collapse}{boolean; whether to collapse the crick strand into watson strand. Default FALSE}

\item{fill}{boolean; flag whether to fill the output matrixes with all CpGs in ref_cpgs. This must be TRUE for HDF5-based experiments. Performance is typically better if fill = TRUE, but may require more memory.}

\item{verbose}{boolean; flag to output messages or not.}
}
\value{
data.table containing vector of all indexed methylation values for the input BED
}
\description{
Parses BED files for methylation values using previously generated index genomic coordinates
}
\details{
Creates an NA-based vector populated with methlylation values from the input BED file in the
respective indexed genomic coordinates
}
\examples{
\dontrun{
#Do Nothing
}
}
