% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_beds.R
\name{read_index}
\alias{read_index}
\title{Parse BED files for unique genomic coordinates}
\usage{
read_index(
  files,
  n_threads = 0,
  batch_size = 200,
  zero_based = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{files}{List of BED files}

\item{batch_size}{Number of files to process before running unique. Default of 30.}
}
\value{
data.table containing all unique genomic coordinates
}
\description{
Parse BED files for unique genomic coordinates
}
\details{
Create list of unique genomic regions from input BED files. Populates a list of batch_size+1 with
the genomic coordinates from BED files, then runs unique() when the list is full and keeps the running
results in the batch_size+1 position. Also indexes based on 'chr' and 'start' for later searching.
}
