% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_beds.R
\name{read_mem_data}
\alias{read_mem_data}
\title{Writes values from input BED files into an in-memory \code{\link{matrix}}}
\usage{
read_mem_data(
  files,
  ref_cpgs,
  col_list,
  batch_size = 20,
  n_threads = 1,
  zero_based = FALSE,
  strand_collapse = FALSE,
  verbose = TRUE,
  fill = TRUE
)
}
\arguments{
\item{files}{list of strings; file.paths of BED files}

\item{ref_cpgs}{data.table; list of CpG sites in the tab-delimited format of chr-start-end. Must be zero-based genome.}

\item{col_list}{The column index object for the input BED files}

\item{batch_size}{integer; Max number of files to hold in memory at once. Default 20}

\item{n_threads}{integer; number of threads to use. Default 1.
Be-careful - there is a linear increase in memory usage with number of threads. This option is does not work with Windows OS.}

\item{zero_based}{boolean; flag for whether the input data is zero-based or not}

\item{strand_collapse}{boolean; whether to collapse the crick strand into watson strand. Default FALSE}

\item{verbose}{boolean; flag to output messages or not.}

\item{fill}{boolean; flag whether to fill the output matrixes with all CpGs in ref_cpgs. This must be TRUE for HDF5-based experiments.}
}
\value{
matrix of the methylation values for input BED files
}
\description{
Writes values from input BED files into an in-memory \code{\link{matrix}}
}
\details{
Using the generated index for genomic coordinates, creates a NA-based dense matrtix of methylation
values for each BED file/sample. Each column contains the meth. values for a single sample.
}
\examples{
\dontrun{
#Do Nothing
}
}
