% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_dimensionality.R
\name{reduce_cpgs}
\alias{reduce_cpgs}
\title{Reduces a assay to a representative matrix}
\usage{
reduce_cpgs(
  scm,
  assay = "score",
  var = c("top", "rand"),
  top_var = 1000,
  na.rm = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{scm}{scMethrix; Input \code{\link{scMethrix}} object}

\item{assay}{string; The assay to use. Default is 'score'}

\item{var}{strning; Choose between random CpG sites ('rand') or most variable CpGs ('top'). Default 'top'}

\item{top_var}{integer; Number of variable CpGs to use. Default 1000 Set it to NULL to use all CpGs (which is not recommended due to memory requirements).}

\item{na.rm}{boolean; flag to remove NA values}

\item{verbose}{boolean; flag to output messages or not}
}
\value{
matrix; the reduced form of the input assay
}
\description{
Reduces a assay to a representative matrix
}
\details{
For the purposes of dimensionality reduction, this function selects either random CpGs or those with the highest variability.
}
\examples{
data('scMethrix_data')
reduce_cpgs(scMethrix_data)
}
