% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_dimensionality.R
\name{reduce_matrix}
\alias{reduce_matrix}
\title{Reduces a matrix}
\usage{
reduce_matrix(
  scm,
  assay = "score",
  var = "top",
  top_var = 1000,
  verbose = FALSE
)
}
\arguments{
\item{scm}{Input \code{\link{scMethrix}} object}

\item{assay}{The assay to use. Default is 'score'}

\item{var}{Choose between random CpG sites ('rand') or most variable CpGs ('top').}

\item{top_var}{Number of variable CpGs to use. Default 1000 Set it to NULL to use all CpGs (which is not recommended due to memory requirements). This option is mutually exclusive with \code{ranges}.}

\item{verbose}{flag to output messages or not}
}
\value{
Reduced matrix
}
\description{
Reduces a matrix
}
\examples{
data('scMethrix_data')
reduce_matrix(scMethrix_data)
}
