% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_dimensionality.R
\name{reduce_scMethrix}
\alias{reduce_scMethrix}
\title{Reduces a assay to a representative matrix}
\usage{
reduce_scMethrix(
  scm,
  assay = "score",
  var = c("top", "random"),
  n_cpg = 1000,
  na.rm = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{assay}{string; name of an existing assay. Default = "score"}

\item{var}{string; Choose between random CpG sites ('rand') or most variable CpGs ('top'). Default 'top'. Seed for random sampling = n_cpg.}

\item{n_cpg}{integer; Number of variable CpGs to reduce to. Default 1000.}

\item{na.rm}{boolean; flag to remove features with any NA values. Final number of CpGs will likely be less than n_cpg. For 'random', Will iterate up to 10 times trying to find random cpgs.}

\item{verbose}{boolean; Flag for outputting function status messages. Default = TRUE}
}
\value{
matrix; the reduced form of the input assay
}
\description{
Reduces a assay to a representative matrix
}
\details{
For the purposes of dimensionality reduction, this function selects either random CpGs or those with the highest variability.
}
\examples{
data('scMethrix_data')
nrow(scMethrix_data)
scMethrix_data <- reduce_scMethrix(scMethrix_data)
nrow(scMethrix_data)
}
