% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_operations.R
\name{region_filter}
\alias{region_filter}
\title{Filter matrices by region}
\usage{
region_filter(m, regions = NULL, type = "any")
}
\arguments{
\item{m}{\code{\link{scMethrix}} object}

\item{regions}{genomic regions to filter-out. Could be a data.table with 3 columns (chr, start, end) or a \code{GenomicRanges} object}

\item{type}{defines the type of the overlap of the CpG sites with the target regions. Default value is \code{any}. For detailed description,
see the \code{foverlaps} function of the \code{\link{data.table}} package.}
}
\value{
An object of class \code{\link{scMthrix}}
}
\description{
Filter matrices by region
}
\details{
Takes \code{\link{scMethrix}} object and filters CpGs based on supplied regions in data.table or GRanges format
}
\examples{
data('scMethrix_data')
# Filter methylation data to chr1:1:5
regions <- GenomicRanges::GRanges(seqnames = "chr1", ranges = IRanges(1,5)) 
region_filter(scMethrix_data$mem, regions)
}
