% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_operations.R
\name{remove_assay}
\alias{remove_assay}
\title{Removes an assay from an \code{\link{scMethrix}} object}
\usage{
remove_assay(scm, assay)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{assay}{string; name of an existing assay. Default = "score"}
}
\value{
An \code{\link{scMethrix}} object
}
\description{
Removes an assay from an \code{\link{scMethrix}} object
}
\examples{
data('scMethrix_data')
remove_assay(scMethrix_data,assay="counts")
}
