% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_operations.R
\name{save_scMethrix}
\alias{save_scMethrix}
\title{Saves an \code{\link{scMethrix}} object}
\usage{
save_scMethrix(
  scm = NULL,
  dest = NULL,
  replace = FALSE,
  quick = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{replace}{Should it overwrite the pre-existing data? FALSE by default.}

\item{quick}{boolean; Flag to skip realizing of matrix operations}

\item{verbose}{boolean; Flag for outputting function status messages. Default = TRUE}

\item{...}{Parameters to pass to saveHDF5SummarizedExperiment}
}
\value{
invisible \code{\link{scMethrix}} object, with the assays stored in the h5_dir
}
\description{
Saves an \code{\link{scMethrix}} object
}
\details{
HDF5 and in-memory \code{\link{scMethrix}} objects are stored differently:
\itemize{
\item HDF5 objects have two files: assays.h5 and se.rds. These files are hardcoded in the saveHDF5SummarizedExperiment function and cannot be changed, as the load functions will only look for these files. To use these, you must specify the directory they will be stored in. That function is somewhat dangerously coded as well, as it will delete everything else in the directory when you save something. Here, a menu prompt has been added to warn the user.
\item In-memory objects are simply stored in a .RDS container. There is no requirement for file name or such, but will still prompt if the file already exists
Using the flag 'replace = TRUE' will override the menus, but care must be taken to not delete important files (this happened numerous times to the authors!).
}

If \code{quick = TRUE} for HDF5 experiments, any operations done on assay matrices will not be realized. In other words, the assay information on the hard disk will not be changed. Non-matrix information will be updated (e.g., metadata) as well as any pending matrix operations. To use this, the experiment must have previously been saved using \code{quick = FALSE}.
}
\examples{
data('scMethrix_data')
dir <- paste0(tempdir(),"/h5")
scm <- convert_scMethrix(scMethrix_data, h5_dir=dir)
save_scMethrix(scm, h5_dir = dir, replace = TRUE)
}
