% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_object.R
\docType{class}
\name{scMethrix-class}
\alias{scMethrix-class}
\alias{scMethrix}
\title{Class scMethrix}
\description{
S4 class scMethrix
}
\section{Slots}{

\describe{
\item{\code{assays}}{A list of two matrices containing 'Methylation' and 'Coverage' information
#slot bins A list of matricies for different size of binning for methylation data}

\item{\code{elementMetadata}}{A DataFrame describing rows in correspoding assay matrices.}

\item{\code{colData}}{genome: the name of the BSgenome that was used to extract CpGs, isHDF5: is it stored in HDF5 Array format}

\item{\code{metadata}}{a list of meta data associated with the assays}

\item{\code{NAMES}}{NULL}

\item{\code{int_colData}}{NULL}

\item{\code{int_metadata}}{NULL}

\item{\code{int_elementMetadata}}{NULL}
}}

