% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_plot.R
\name{scMethrix_theme}
\alias{scMethrix_theme}
\title{Theme for ggplot}
\usage{
scMethrix_theme(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{Size of text}

\item{base_family}{Family of text}
}
\value{
ggplot element; data for the ggplot theme
}
\description{
Theme for ggplot
}
