% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_funcs.R
\name{split_granges}
\alias{split_granges}
\title{Chunks a \code{\link{GRanges}} object by \code{factor}, \code{percent} or \code{number}}
\usage{
split_granges(gr, factor = NA, percent = NA, num = NA)
}
\arguments{
\item{gr}{The \code{\link{GRanges}} object}

\item{factor}{The factor to which divide the chunks into}

\item{percent}{The size in percentage of CpGs in each chunk}

\item{num}{The number of CpGs to include in each chunk}
}
\value{
\code{\link{GRangesList}} containing all the chunked \code{\link{GRanges}}
}
\description{
Chunks a \code{\link{GRanges}} object by \code{factor}, \code{percent} or \code{number}
}
\details{
Divides \code{\link{GRanges}} into a list based on a chunking parameter
}
\examples{
regions <- GenomicRanges::GRanges(seqnames = "chr1", ranges = IRanges(1,100))
regions <- bin_granges(regions,bin_size=10)
split_granges(regions, num=10)
split_granges(regions, factor=10)
split_granges(regions, percent=10)
}
