% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_funcs.R
\name{split_time}
\alias{split_time}
\title{Outputs the split/lap/iteration time}
\usage{
split_time()
}
\value{
Returns formatted elapsed time since start_time or last split_time
}
\description{
Outputs the split/lap/iteration time
}
\details{
Gets the stored elapsed proc.time() from either the initial start_time or the previous split_time
}
