% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_funcs.R
\name{split_vector}
\alias{split_vector}
\title{Splits a vector into subvectors}
\usage{
split_vector(vec, num, by = c("chunk", "size"))
}
\arguments{
\item{vec}{The vector to split}

\item{num}{The number to split by}

\item{by}{Whether to split by chunk or by size}
}
\value{
A vector of vectors
}
\description{
Splits a vector into subvectors
}
\details{
Splits a vector into \code{num} vectors or vectors of \code{n} size. If \code{len(vec)\%\%num != 0},
then the last vector will have a length of less than size \code{num}
}
\examples{
# Split vector into 4 sub vectors
split_vector(c(1,2,3,4,5,6,7,8),4,by="chunk")

# Split vector into sub-vectors with a size of 2
split_vector(c(1,2,3,4,5,6,7,8),2,by="size")
}
