% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_funcs.R
\name{start_time}
\alias{start_time}
\title{Starts an internal stopwatch}
\usage{
start_time()
}
\description{
Starts an internal stopwatch
}
\details{
Save the current time to later use for split/lap and overall times
}
