% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_funcs.R
\name{subset_ref_cpgs}
\alias{subset_ref_cpgs}
\title{Subsets a given list of CpGs by another list of CpGs}
\usage{
subset_ref_cpgs(ref_cpgs, gen_cpgs, verbose = TRUE)
}
\arguments{
\item{ref_cpgs}{A reference set of CpG sites (e.g. Hg19 or mm10) in bedgraph format}

\item{gen_cpgs}{A subset of CpG sites. Usually obtained from \code{\link{read_index}}.}

\item{verbose}{flag to output messages or not}
}
\value{
Returns list of CpG sites in bedgraph format
}
\description{
Subsets a given list of CpGs by another list of CpGs
}
\details{
Typically used to reduce the number of potential CpG sites to include only those present
in the input files so as to maximize performance and minimize resources. Can also be used for quality
control to see if there is excessive number of CpG sites that are not present in the reference genome.
}
\examples{
ref_cpgs = data.frame(chr="chr1",start=(1:5*2-1), end=(1:5*2))
subset_ref_cpgs(ref_cpgs,ref_cpgs[1:3,])
}
