% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_operations.R
\name{transform_assay}
\alias{transform_assay}
\title{Transforms an assay in an \code{\link{scMethrix}} object.}
\usage{
transform_assay(m, assay = NULL, name = NULL, trans = NULL)
}
\arguments{
\item{m}{A \code{\link{scMethrix}} object}

\item{assay}{A \code{\link{scMethrix}} object}

\item{name}{Merge by columns or rows}

\item{trans}{The transformation function}
}
\value{
An \code{\link{scMethrix}} object
}
\description{
Transforms an assay in an \code{\link{scMethrix}} object.
}
\examples{
data('scMethrix_data')
#merge_scMethrix(scMethrix_data$mem,scMethrix_data$mem)
}
