% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_transforms.R
\name{umap_scMethrix}
\alias{umap_scMethrix}
\title{Generates UMAP for scMethrix}
\usage{
umap_scMethrix(scm)
}
\arguments{
\item{scm}{A \code{\link{scMethrix}} object}
}
\value{
An \code{\link{scMethrix}} object
}
\description{
Generates UMAP for scMethrix
}
\details{
Does UMAP stuff
}
\examples{
data('scMethrix_data')
}
