% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_ops.R
\name{bin_granges}
\alias{bin_granges}
\title{Bins each region in a \code{\link{GRanges}} object into bins of specified \code{bin_size}}
\usage{
bin_granges(gr, bin_size = 1e+05)
}
\arguments{
\item{gr}{GRanges; The \code{\link{GRanges}} object}

\item{bin_size}{integer; x > 0; The length of region in each bin}
}
\value{
\code{\link{GRangesList}} containing all the binned \code{\link{GRanges}}
}
\description{
Bins each region in a \code{\link{GRanges}} object into bins of specified \code{bin_size}
}
\details{
Bins a single region in \code{\link{GRanges}} format into multiple regions with a specified \code{bin_size}. If \code{length(gr) \%\% bin_size != 0}, then the last GRange will have a length < \code{bin_size}. This is used instead of tile when you need consistently sized bins with the last bin being smaller
}
\examples{
regions <- GenomicRanges::GRanges(seqnames = "chr1", ranges = IRanges(1,10000))
bin_granges(regions,bin_size=1000) 
}
