% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_funcs.R
\name{binarize}
\alias{binarize}
\title{Binarize an input}
\usage{
binarize(x, val = 50)
}
\arguments{
\item{x}{A value to binarize}

\item{val}{The threshold for binarizing}
}
\value{
1 or 0, if above of below the threshold
}
\description{
Binarize an input
}
\examples{
binarize(0.7, val=0.5)
}
