% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_funcs.R
\name{binarize}
\alias{binarize}
\title{Binarize an input value based on a \code{threshold}}
\usage{
binarize(x, threshold = 50)
}
\arguments{
\item{x}{A value to binarize}

\item{threshold}{The threshold for binarizing}
}
\value{
1 or 0, if above of below the threshold
}
\description{
Binarize an input value based on a \code{threshold}
}
\examples{
vals <- c(0,0.25,0.5,0.75,1)
binarize(vals, threshold=0.5)
}
