% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_funcs.R
\name{cast_granges}
\alias{cast_granges}
\title{Casts genomic regions into \code{\link{GRanges}} format}
\usage{
cast_granges(regions)
}
\arguments{
\item{regions}{The input regions}
}
\value{
\code{\link{GRanges}} object with the input regions
}
\description{
Casts genomic regions into \code{\link{GRanges}} format
}
\details{
Casts the input as a \code{\link{GRanges}} object. Input can be \code{\link{GRanges}} or a
\code{\link{data.frame}}-compatible class that can be cast through \code{as.data.frame()}. Input BED format
must be \code{chr-start-end} for \code{\link{data.frame}} objects.
}
\examples{
regions = data.table(chr = 'chr1', start = 1, end = 100)
cast_granges(regions) 
}
