% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_operations.R
\name{convert_HDF5_scMethrix}
\alias{convert_HDF5_scMethrix}
\title{Converts HDF5 \code{\link{scMethrix}} object to standard in-memory object.}
\usage{
convert_HDF5_scMethrix(m = NULL)
}
\arguments{
\item{m}{An object of class \code{\link{scMethrix}}, HDF5 format}
}
\value{
An object of class \code{\link{scMethrix}}
}
\description{
Converts HDF5 \code{\link{scMethrix}} object to standard in-memory object.
}
\details{
Takes a \code{\link{scMethrix}} object and returns with the same object with in-memory assay slots.
}
\examples{
data('scMethrix_data')
dir <- paste0(tempdir(),"/h5")
m <- convert_scMethrix(scMethrix_data, h5_dir=dir)
convert_HDF5_scMethrix(m)
}
