% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_operations.R
\name{convert_methrix}
\alias{convert_methrix}
\title{Converts an in-memory object to an on-disk HDF5 object.}
\usage{
convert_methrix(m = NULL, h5_dir = NULL)
}
\arguments{
\item{m}{An object of class \code{\link{scMethrix}}}
}
\value{
An object of class \code{\link{scMethrix}}, HDF5 format
}
\description{
Converts an in-memory object to an on-disk HDF5 object.
}
\details{
Takes a \code{\link{scMethrix}} object and returns with the same object with delayed array assay slots
with HDF5 backend. Might take long time!
}
