% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_dimensionality.R
\name{dim_red_scMethrix}
\alias{dim_red_scMethrix}
\title{Reduces dimensionality (tSNE, UMAP, PCA, or custom)}
\usage{
dim_red_scMethrix(
  scm,
  assay = "score",
  type = c("tSNE", "UMAP", "PCA"),
  perplexity = 30,
  verbose = FALSE,
  n_components = 2,
  n_neighbors = 15,
  ...
)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{assay}{string; name of an existing assay. Default = "score"}

\item{type}{string; the type of imputation "tSNE","UMAP", or "PCA"}

\item{perplexity}{numeric; Perplexity parameter (should not be bigger than 3 * perplexity < nrow(X) - 1, see details for interpretation)}

\item{verbose}{boolean; Flag for outputting function status messages. Default = TRUE}

\item{n_components}{integer; Number of components to use}

\item{n_neighbors}{integer; number of nearest neighbors for UMAP}

\item{...}{Other arguments that can be passed to Rtsne}
}
\value{
\code{\link{scMethrix}} object with reducedDim assay
}
\description{
Reduces dimensionality (tSNE, UMAP, PCA, or custom)
}
\details{
Does reduction stuff
}
\examples{
data('scMethrix_data')

}
\seealso{
\code{\link[=plot_dim_red]{plot_dim_red()}} for plotting
}
