% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_dimensionality.R
\name{dim_red_scMethrix}
\alias{dim_red_scMethrix}
\title{Reduces dimensionality}
\usage{
dim_red_scMethrix(
  scm,
  assay = "score",
  type = c("tSNE", "UMAP", "PCA"),
  var = "top",
  top_var = 1000,
  perplexity = 30,
  verbose = FALSE,
  n_components = 2,
  n_neighbors = 15,
  ...
)
}
\arguments{
\item{scm}{scMethrix; Input \code{\link{scMethrix}} object}

\item{assay}{string; The assay to use. Default is 'score'}

\item{type}{string; the type of imputation "tSNE","UMAP", or "PCA"}

\item{var}{strning; Choose between random CpG sites ('rand') or most variable CpGs ('top'). Default 'top'}

\item{top_var}{integer; Number of variable CpGs to use. Default 1000 Set it to NULL to use all CpGs (which is not recommended due to memory requirements).}

\item{perplexity}{numeric; Perplexity parameter (should not be bigger than 3 * perplexity < nrow(X) - 1, see details for interpretation)}

\item{verbose}{boolean; flag to output messages or not}

\item{n_components}{integer; Number of components to use for PCA}

\item{n_neighbors}{integer; number of nearest neighbors}

\item{...}{Other arguments that can be passed to Rtsne}
}
\value{
\code{\link{scMethrix}} object with reducedDim assay
}
\description{
Reduces dimensionality
}
\details{
Does reduction stuff
}
\examples{
data('scMethrix_data')

}
\seealso{
\code{\link[=plot_dim_red]{plot_dim_red()}} for plotting
}
