% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs.R
\name{.validateArg}
\alias{.validateArg}
\title{Validates arguments. Allows partial matching.}
\usage{
.validateArg(arg, parent = NULL, ignore.case = T)
}
\arguments{
\item{arg}{variable; the variable in which to check}

\item{parent}{closure; the parent function in which to check the input}

\item{ignore.case}{boolean; ignores case of the choices}
}
\value{
arg, if the value is in the function definition.
}
\description{
Validates arguments. Allows partial matching.
}
\details{
Check the parent function input arguments to see whether the inputted value is part of the set. Will return a formatted error message with the incorrect variable name and all the acceptable inputs.

To use, it can be called as such:

genericFunc <- function(values = c("apple","orange","banana")) {

values <- .validateArg(values)

}

If the argument for values is acceptable (e.g. "apple" or "ban"), it will return the matched string.

This can be used with pipes, but will give erroneous names for the input variable
}
