% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs.R
\name{.validateAssay}
\alias{.validateAssay}
\title{Validates an assay is in the object. Allows partial pattern.}
\usage{
.validateAssay(scm = NULL, assay = NULL, check.absent = F)
}
\arguments{
\item{scm}{scMethrix; the experiment object}

\item{assay}{string; the name of the assay}

\item{check.absent}{boolean; Checks if the assay is present}
}
\value{
string or boolean; if \code{check.absent == T}, the name of the matched assay or error if it doesn't exist. If \code{check_assay == F}, the boolean value for if the assay exists in the experiment
}
\description{
Validates an assay is in the object. Allows partial pattern.
}
\details{
Check the assays in an scMethrix object and partial matches
}
