% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs.R
\name{.validateExp}
\alias{.validateExp}
\title{Validates to see if object is a proper scMethrix object}
\usage{
.validateExp(scm)
}
\arguments{
\item{scm}{scMethrix; the experiment object to test}
}
\value{
invisible(TRUE), if the object is valid. Error if not.
}
\description{
Validates to see if object is a proper scMethrix object
}
