% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_transforms.R
\name{generate_training_set}
\alias{generate_training_set}
\title{Splits an experiment into two for use as a training and test set}
\usage{
generate_training_set(scm = NULL, training_prop = 0.2)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{training_prop}{numeric; The size of the training set as a proportion of the experiment (0 to 1)
For a range, the optimal value will be estimated; this is time-intensive.}
}
\value{
list; two \code{\link{scMethrix}} objects names 'training' and 'test'
}
\description{
Splits an experiment into two for use as a training and test set
}
\details{
Does stuff
}
\examples{
data('scMethrix_data')
generate_training_set(scMethrix_data, training_prop = 0.2)
}
