% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_object.R
\name{generic_scMethrix_function}
\alias{generic_scMethrix_function}
\title{Function used only for inheritence for Roxygen2 documentation. Lists the common function inputs used in the package}
\usage{
generic_scMethrix_function(
  scm,
  assay,
  new_assay,
  trans,
  verbose,
  n_chunks,
  n_threads,
  h5_dir,
  overlap_type,
  batch_size,
  replace,
  na.rm
)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{assay}{string; name of an existing assay. Default = "score"}

\item{new_assay}{string; name for transformed assay. Default = "new_assay"}

\item{trans}{closure; The transformation function. Default = mean}

\item{verbose}{boolean; Flag for outputting function status messages. Default = TRUE}

\item{n_chunks}{integer; Number of chunks to split the \code{\link{scMethrix}} object in case it is very large. Default = 1}

\item{n_threads}{integer; Maximum number of parallel instances. Default = 1}

\item{h5_dir}{string; The directory to store HDF5 files. Will be created if it does not exist. Default = NULL}

\item{overlap_type}{defines the type of the overlap of the CpG sites with the target region. Default value is \code{within}. For detailed description, see the \code{findOverlaps} function of the \code{\link{IRanges}} package.}

\item{batch_size}{integer; The maximum number of elements to process at once.}

\item{replace}{boolean; flag for whether to delete the contents of \code{h5_dir} before saving}

\item{na.rm}{boolean; flag to remove NA values}
}
\description{
Function used only for inheritence for Roxygen2 documentation. Lists the common function inputs used in the package
}
