% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_operations.R
\name{get_matrix}
\alias{get_matrix}
\title{Extract methylation or coverage matrices}
\usage{
get_matrix(m = NULL, add_loci = FALSE, in_granges = FALSE, type = "score")
}
\arguments{
\item{m}{\code{\link{scMethrix}} object}

\item{add_loci}{Default FALSE. If TRUE adds CpG position info to the matrix and returns as a data.table}

\item{in_granges}{Do you want the outcome in \code{\link{GRanges}}?}

\item{type}{Which matrix to get, "m": methlation, "c": coverage}
}
\value{
Coverage or Methylation matrix
}
\description{
Extract methylation or coverage matrices
}
\details{
Takes \code{\link{scMethrix}} object and returns the \code{methylation} matrix
}
\examples{
data('scMethrix_data')

# Get methylation data
get_matrix(scMethrix_data$mem)

# Get methylation data with loci
get_matrix(scMethrix_data$mem, add_loci=TRUE)

# Get methylation data with loci inside a Granges object 
get_matrix(scMethrix_data$mem, add_loci=TRUE, in_granges=TRUE)
}
