% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_operations.R
\name{get_metadata_stats}
\alias{get_metadata_stats}
\title{Adds descriptive statistics to metadata columns in an \code{\link{scMethrix}} object.}
\usage{
get_metadata_stats(scm)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}
}
\value{
An \code{\link{scMethrix}} object
}
\description{
Adds descriptive statistics to metadata columns in an \code{\link{scMethrix}} object.
}
\details{
Adds the mean, median, SD, and sample count and coverage (if present) for
the \code{\link{GenomicRanges}} in an \code{\link{scMethrix}} object. This can be accessed using mcols().
}
\examples{
data('scMethrix_data')
get_metadata_stats(scMethrix_data)
}
