% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_funcs.R
\name{get_sample_name}
\alias{get_sample_name}
\title{Returns sample name derived from the input file name}
\usage{
get_sample_name(s)
}
\arguments{
\item{s}{string; A file.path}
}
\value{
string containing the sample name
}
\description{
Returns sample name derived from the input file name
}
\examples{
#get_sample_name("C:/dir/dir/filename.ext")
}
