% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_funcs.R
\name{get_sample_name}
\alias{get_sample_name}
\title{Returns file name minus the extension from a file.path to represent the sample name}
\usage{
get_sample_name(s)
}
\arguments{
\item{s}{A file.path}
}
\value{
string containing the sample name
}
\description{
Returns file name minus the extension from a file.path to represent the sample name
}
\details{
Acts in the same way as ternary operators in other language. Must surround with brackets to use)
}
