% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_operations.R
\name{get_stats}
\alias{get_stats}
\title{Estimate descriptive statistics}
\usage{
get_stats(m, per_chr = TRUE)
}
\arguments{
\item{m}{\code{\link{scMethrix}} object}

\item{per_chr}{Estimate stats per chromosome. Default TRUE}
}
\value{
data.table of summary stats
}
\description{
Estimate descriptive statistics
}
\details{
Calculate descriptive statistics
}
\examples{
data('scMethrix_data')

#Get stats for each sample and chromosome
get_stats(scMethrix_data$mem)

#Get stats for each sample
get_stats(scMethrix_data$mem,per_chr = FALSE)
}
\seealso{
\code{\link{plot_stats}}
}
