% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_operations.R
\name{get_stats}
\alias{get_stats}
\title{Estimate descriptive statistics for each sample}
\usage{
get_stats(scm = NULL, per_chr = TRUE)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{per_chr}{boolean; Estimate stats per chromosome. Default TRUE}
}
\value{
data.table of summary stats
}
\description{
Estimate descriptive statistics for each sample
}
\details{
Calculate descriptive statistics (mean, median, SD) either by sample or \code{per_chr}
}
\examples{
data('scMethrix_data')

#Get stats for each sample and chromosome
get_stats(scMethrix_data)

#Get stats for each sample
get_stats(scMethrix_data,per_chr = FALSE)
}
