% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_transforms.R
\name{impute_by_iPCA}
\alias{impute_by_iPCA}
\title{Imputes the an \code{\link{scMethrix}}object using a iterative PCA model}
\usage{
impute_by_iPCA(
  scm = NULL,
  assay = "score",
  new_assay = "impute",
  n_pc = 2,
  ...
)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{assay}{string; name of an existing assay. Default = "score"}

\item{new_assay}{string; name for transformed assay. Default = "new_assay"}

\item{n_pc}{integer > 0; The number of principal components to use. This can be a range - e.g. c(1,5) - or a single value.
For a range, the optimal value will be estimated; this is time-intensive.}

\item{...}{Additional arguments for \link[missMDA]{imputePCA}}
}
\value{
An \code{\link{scMethrix}} object
}
\description{
Imputes the an \code{\link{scMethrix}}object using a iterative PCA model
}
\details{
Uses the inputted function to transform an assay in the \code{\link{scMethrix}} object.
}
\examples{
data('scMethrix_data')
impute_by_iPCA(scMethrix_data, assay = "score", new_assay = "impute", n_pc = 2)
}
\references{
Bro, R., Kjeldahl, K. Smilde, A. K. and Kiers, H. A. L. (2008) Cross-validation of component models: A critical look at current methods. Analytical and Bioanalytical Chemistry, 5, 1241-1251.

Josse, J. and Husson, F. (2011). Selecting the number of components in PCA using cross-validation approximations. Computational Statistics and Data Analysis. 56 (6), pp. 1869-1879.
}
