% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_transforms.R
\name{impute_by_kNN}
\alias{impute_by_kNN}
\title{Imputes the an \code{\link{scMethrix}}object using a random forest model}
\usage{
impute_by_kNN(scm = NULL, assay = "score", new_assay = "impute", k = 10, ...)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{assay}{string; name of an existing assay. Default = "score"}

\item{new_assay}{string; name for transformed assay. Default = "new_assay"}

\item{k}{Number of neighbors to be used in the
    imputation (default=10)}

\item{...}{Additional arguments for missForest::missForest}
}
\value{
An \code{\link{scMethrix}} object
}
\description{
Imputes the an \code{\link{scMethrix}}object using a random forest model
}
\details{
Uses the inputted function to transform an assay in the \code{\link{scMethrix}} object.
}
\examples{
data('scMethrix_data')
impute_by_RF(scMethrix_data, assay = "score", new_assay = "impute")
}
