% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_transforms.R
\name{impute_by_melissa}
\alias{impute_by_melissa}
\title{Imputes the NA values of a \code{\link{scMethrix}} object.}
\usage{
impute_by_melissa(scm, threshold = 50, assay = "score", new_assay = "impute")
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{threshold}{The value for cutoff in the "score" assay to determine methylated or unmethylated status.
Default = 50}

\item{assay}{string; name of an existing assay. Default = "score"}

\item{new_assay}{string; name for transformed assay. Default = "new_assay"}
}
\value{
An \code{\link{scMethrix}} object
}
\description{
Imputes the NA values of a \code{\link{scMethrix}} object.
}
\details{
Uses the inputted function to transform an assay in the \code{\link{scMethrix}} object
}
\examples{
data('scMethrix_data')
}
\references{
Kapourani CA, Sanguinetti G (2019). “Melissa: Bayesian clustering and imputation of single cell methylomes.” Genome Biology, 20, 61. doi: 10.1186/s13059-019-1665-8.
}
