% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_transforms.R
\name{impute_regions}
\alias{impute_regions}
\title{Generic imputation return function}
\usage{
impute_regions(
  scm = NULL,
  assay = "score",
  new_assay = "impute",
  regions = NULL,
  n_chunks = 1,
  n_threads = 1,
  overlap_type = "within",
  type = c("kNN", "iPCA", "RF"),
  verbose = TRUE,
  k = 10,
  n_pc = 2,
  ...
)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{assay}{string; name of an existing assay. Default = "score"}

\item{new_assay}{string; name for transformed assay. Default = "new_assay"}

\item{regions}{Granges; the regions to impute. Default is by chromosome.}

\item{n_chunks}{integer; Number of chunks to split the \code{\link{scMethrix}} object in case it is very large. Default = 1}

\item{n_threads}{integer; Maximum number of parallel instances. Default = 1}

\item{overlap_type}{defines the type of the overlap of the CpG sites with the target region. Default value is \code{within}. For detailed description, see the \code{findOverlaps} function of the \code{\link{IRanges}} package.}

\item{type}{string/closure; the imputation to perform "kNN","iPCA",or "RF". Otherwise, a closure can be specified that returns the imputed matrix. Default = "kNN"}

\item{verbose}{boolean; Flag for outputting function status messages. Default = TRUE}

\item{k}{Number of neighbors to be used in the
    imputation (default=10)}

\item{n_pc}{the range of principal components to check when using iPCA. Caution: this can be very time-intensive}

\item{...}{further arguments passed to or from other methods}
}
\value{
list; two \code{\link{scMethrix}} objects names 'training' and 'test'
}
\description{
Generic imputation return function
}
\details{
Uses the specified imputation operation to evaluation an scMethrix object.
}
\examples{
data('scMethrix_data')
impute_regions(scMethrix_data)
}
\references{
Hastie T, Tibshirani R, Narasimhan B, Chu G (2021). impute: impute: Imputation for microarray data. R package version 1.66.0.

Stekhoven, D. J., & Bühlmann, P. (2012). MissForest—non-parametric missing value imputation for mixed-type data. Bioinformatics, 28(1), 112-118.

Bro, R., Kjeldahl, K. Smilde, A. K. and Kiers, H. A. L. (2008) Cross-validation of component models: A critical look at current methods. Analytical and Bioanalytical Chemistry, 5, 1241-1251.

Josse, J. and Husson, F. (2011). Selecting the number of components in PCA using cross-validation approximations. Computational Statistics and Data Analysis. 56 (6), pp. 1869-1879.
}
