% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_funcs.R
\name{is_h5}
\alias{is_h5}
\title{Checks if \code{\link{scMethrix}} object is an HDF5 object}
\usage{
is_h5(scm)
}
\arguments{
\item{scm}{scMethrix; The \code{\link{scMethrix}} object}
}
\value{
boolean; Whether the object is HDF5
}
\description{
Checks if \code{\link{scMethrix}} object is an HDF5 object
}
\details{
This checks the metadata whether the experiment is in HDF5 format. As this can be manually changed and will
not update after user operations to the assays, it may not be reliable.
may not be reliable
}
\examples{
data('scMethrix_data')
is_h5(scMethrix_data)
}
