% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_operations.R
\name{mask_helper}
\alias{mask_helper}
\title{Helper function for masking. All rows in row_idx will be set to NA.}
\usage{
mask_helper(scm, row_idx, verbose = TRUE)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{row_idx}{numeric; A vector of row indexes for which to replace with NA}

\item{verbose}{boolean; Flag for outputting function status messages. Default = TRUE}
}
\value{
An object of class \code{\link{scMethrix}}
}
\description{
Helper function for masking. All rows in row_idx will be set to NA.
}
\details{
Takes \code{\link{scMethrix}} object and masks sites with too high or too low coverage
by putting NA for assay values. The sites will remain in the object and all assays will be affected.

\code{low_threshold} is used to mask sites with low overall cell counts. A site represented by a single sample is typically not useful.
\code{prop_threshold} is used to mask sites with a low proportional count
}
