% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_operations.R
\name{merge_scMethrix}
\alias{merge_scMethrix}
\title{Merges two \code{\link{scMethrix}} objects by \code{row} or \code{col}}
\usage{
merge_scMethrix(scm1 = NULL, scm2 = NULL, by = c("row", "col"))
}
\arguments{
\item{scm1}{A \code{\link{scMethrix}} object}

\item{scm2}{A \code{\link{scMethrix}} object}

\item{by}{Merge by columns or rows}
}
\value{
A merged \code{\link{scMethrix}} object
}
\description{
Merges two \code{\link{scMethrix}} objects by \code{row} or \code{col}
}
\details{
Merges the base assay data from two \code{\link{scMethrix}} objects. Merging of additional slot
data is not supported at this time. Non-common assays between objects will be dropped
}
\examples{
data('scMethrix_data')
merge_scMethrix(scMethrix_data[1:5],scMethrix_data[6:10],by="row")
}
