% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_operations.R
\name{merge_scMethrix}
\alias{merge_scMethrix}
\title{Merges two \code{\link{scMethrix}} objects.}
\usage{
merge_scMethrix(m1 = NULL, m2 = NULL, by = c("row", "col"))
}
\arguments{
\item{m1}{A \code{\link{scMethrix}} object}

\item{m2}{A \code{\link{scMethrix}} object}

\item{by}{Merge by columns or rows}
}
\value{
A merged \code{\link{scMethrix}} object
}
\description{
Merges two \code{\link{scMethrix}} objects.
}
\details{
Merges the base assay data from two \code{\link{scMethrix}} objects. Merging of additional slot
data is not supported at this time. Non-common assays between objects will be dropped
}
\examples{
data('scMethrix_data')
#merge_scMethrix(scMethrix_data$mem,scMethrix_data$mem)
}
