% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_operations.R
\name{merge_scMethrix}
\alias{merge_scMethrix}
\title{Merges two \code{\link{scMethrix}} objects by \code{row} or \code{col}}
\usage{
merge_scMethrix(
  scm1 = NULL,
  scm2 = NULL,
  h5_dir = NULL,
  by = c("row", "column"),
  verbose = TRUE
)
}
\arguments{
\item{scm1}{\code{\link{scMethrix}}; A single cell methylation experiment}

\item{scm2}{\code{\link{scMethrix}}; A single cell methylation experiment}

\item{h5_dir}{string; The directory to use. Will be created if it does not exist. Default = NULL}

\item{by}{string; Merge by 'columns' or 'rows'}

\item{verbose}{boolean; Flag for outputting function status messages. Default = TRUE}
}
\value{
A merged \code{\link{scMethrix}} object
}
\description{
Merges two \code{\link{scMethrix}} objects by \code{row} or \code{col}
}
\details{
Merges the assay data from two \code{\link{scMethrix}} objects. Assays not shared between assays will be dropped, as well as all reduced dimensionality data.

Requirements for merging
\itemize{
\item If merging by rows, all CpG sites must be unique and samples must be identical
\item If merging by columns, all samples must be unique and CpG sites must be identical
}

Metadata will be retained in certain situations:
For row merges:
- Ranges metadata (\code{mcols()}) will be merged, with missing columns in either assay filled with NAs
- Sample metadata (\code{colData()}) will attempt to be merged. Overlapping, non-identical columns will be appended with \code{.1} and \code{.2}.

For row merges:
- Ranges metadata (\code{mcols()}) will attempt to be merged. Overlapping, non-identical columns will be appended with \code{.1} and \code{.2}.
- Sample metadata (\code{colData()}) will be merged, with missing columns in either assay filled with NAs

For both merges:
- Experiment metadata (\code{metadata()}) will attempt to be merged. Overlapping, non-identical elements will be appended with \code{.1} and \code{.2}.

Custom experiment metadata can manually be added via \code{metadata() <-}, or to rowRanges via \code{mcols() <-}.
}
\examples{
data('scMethrix_data')
merge_scMethrix(scMethrix_data[1:5],scMethrix_data[6:10],by="row")
merge_scMethrix(scMethrix_data[,1:2],scMethrix_data[,3:4],by="col")
}
