% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_funcs.R
\name{parse_source_idx}
\alias{parse_source_idx}
\title{Subsets a given list of CpGs by another list of CpGs}
\usage{
parse_source_idx(
  chr_idx = NULL,
  start_idx = NULL,
  end_idx = NULL,
  strand_idx = NULL,
  beta_idx = NULL,
  M_idx = NULL,
  U_idx = NULL,
  cov_idx = NULL,
  beta_fract = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{chr_idx}{integer; column of the chromosome}

\item{start_idx}{integer; column of the CpG start site}

\item{end_idx}{integer; column of the CpG end site}

\item{strand_idx}{integer; column of the strand}

\item{beta_idx}{integer; column of the beta value}

\item{M_idx}{integer; column of the # of methylated reads}

\item{U_idx}{integer; column of the # of unmethylated reads}

\item{cov_idx}{integer; column of the coverage}

\item{beta_fract}{double; fraction of methylated bases TODO: check this}

\item{verbose}{flag to output messages or not}
}
\value{
List of column names and indexes
}
\description{
Subsets a given list of CpGs by another list of CpGs
}
\details{
Typically used to reduce the number of potential CpG sites to include only those present  in the input files so as to maximize performance and minimize resources. Can also be used for quality control to see if there is excessive number of CpG sites that are not present in the reference genome.
}
