% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_funcs.R
\name{parse_source_idx}
\alias{parse_source_idx}
\title{Generates the column structure for importing bedgraph files}
\usage{
parse_source_idx(
  chr_idx = NULL,
  start_idx = NULL,
  end_idx = NULL,
  strand_idx = NULL,
  beta_idx = NULL,
  M_idx = NULL,
  U_idx = NULL,
  cov_idx = NULL,
  verbose = TRUE
)
}
\arguments{
\item{chr_idx}{integer; column of the chromosome}

\item{start_idx}{integer; column of the CpG start site}

\item{end_idx}{integer; column of the CpG end site}

\item{strand_idx}{integer; column of the strand}

\item{beta_idx}{integer; column of the beta value}

\item{M_idx}{integer; column of the # of methylated reads}

\item{U_idx}{integer; column of the # of unmethylated reads}

\item{cov_idx}{integer; column of the coverage}

\item{verbose}{flag to output messages or not}
}
\value{
List of column names and indexes
}
\description{
Generates the column structure for importing bedgraph files
}
