% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_dimensionality.R
\name{pca_scMethrix}
\alias{pca_scMethrix}
\title{Principal Component Analysis}
\usage{
pca_scMethrix(scm, assay = "score", verbose = FALSE, ...)
}
\arguments{
\item{scm}{Input \code{\link{scMethrix}} object}

\item{assay}{The assay to use. Default is 'score'}

\item{verbose}{flag to output messages or not}

\item{...}{arguments passed to or from other methods.  If \code{x} is
    a formula one might specify \code{scale.} or \code{tol}.}
}
\value{
\code{\link{scMethrix}} object with reducedDim 'PCA'
}
\description{
Principal Component Analysis
}
\details{
Do PCA stuff
}
\examples{
data('scMethrix_data')
pca_scMethrix(scMethrix_data)
}
