% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_dimensionality.R
\name{pca_scMethrix}
\alias{pca_scMethrix}
\title{Principal Component Analysis}
\usage{
pca_scMethrix(
  scm,
  assay = "score",
  var = "top",
  top_var = 1000,
  verbose = FALSE,
  n_components = 2,
  ...
)
}
\arguments{
\item{scm}{Input \code{\link{scMethrix}} object}

\item{assay}{The assay to use. Default is 'score'}

\item{var}{Choose between random CpG sites ('rand') or most variable CpGs ('top'). Default 'top'}

\item{top_var}{Number of variable CpGs to use. Default 1000 Set it to NULL to use all CpGs (which is not recommended due to memory requirements).}

\item{verbose}{flag to output messages or not}

\item{n_components}{Number of components to use}

\item{...}{arguments passed to or from other methods.  If \code{x} is
    a formula one might specify \code{scale.} or \code{tol}.}
}
\value{
\code{\link{scMethrix}} object with reducedDim 'PCA'
}
\description{
Principal Component Analysis
}
\details{
Do PCA stuff. NA values will be removed.
}
\examples{
data('scMethrix_data')
pca_scMethrix(scMethrix_data)
}
\seealso{
\code{\link[=plot_pca]{plot_pca()}} for plotting
}
