% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_plot.R
\name{plot_pca}
\alias{plot_pca}
\title{Plot PCA results}
\usage{
plot_pca(
  scm = NULL,
  col_anno = NULL,
  shape_anno = NULL,
  pc_x = "PC1",
  pc_y = "PC2",
  show_labels = FALSE,
  plot_vars = FALSE
)
}
\arguments{
\item{scm}{optional scMethrix object. Default NULL}

\item{col_anno}{Column name of colData(m). Default NULL. Will be used as a factor to color different groups. Required \code{methrix} object}

\item{shape_anno}{Column name of colData(m). Default NULL. Will be used as a factor to shape different groups. Required \code{methrix} object}

\item{pc_x}{Default 'PC1'}

\item{pc_y}{Default 'PC2'}

\item{show_labels}{Default FLASE}

\item{plot_vars}{Create additional plot for principal component shares}
}
\value{
ggplot2 object
}
\description{
Plot PCA results
}
\examples{
data('scMethrix_data')
scmpc = pca_scMethrix(scMethrix_data)
plot_pca(scmpc)
}
