% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_plot.R
\name{plot_sparsity}
\alias{plot_sparsity}
\title{Sparsity of sample
inheritParams generic_plot_function}
\usage{
plot_sparsity(
  scm = NULL,
  assay = "score",
  type = c("box", "scatter", "both"),
  pheno = NULL,
  show_legend = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{assay}{string; name of an existing assay. Default = "score"}

\item{type}{string; Choose between 'box' (boxplot) or 'scatter' (scatterplot).}

\item{pheno}{string; Col name of colData(m). Will be used as a factor to color different groups}

\item{show_legend}{boolean; Display the legend on the plot}

\item{verbose}{boolean; Flag for outputting function status messages. Default = TRUE}

\item{...}{Additional parameters to feed to scMethrix_theme()}
}
\value{
ggplot2 object
}
\description{
Sparsity of sample
inheritParams generic_plot_function
}
\examples{
data('scMethrix_data')
plot_sparsity(scm = scMethrix_data)
}
