% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_plot.R
\name{plot_stats}
\alias{plot_stats}
\title{Plot descriptive statistics}
\usage{
plot_stats(
  scm,
  assay = "score",
  stat = c("mean", "median", "count"),
  per_chr = FALSE,
  ignore_chr = NULL,
  ignore_samples = NULL,
  n_col = NULL,
  n_row = NULL,
  pheno = NULL,
  verbose = TRUE,
  show_legend = FALSE,
  ...
)
}
\arguments{
\item{scm}{scMethrix; \code{\link{get_stats}} will be run for the specified assay}

\item{assay}{string; name of an existing assay. Default = "score"}

\item{stat}{string; Can be \code{mean} or \code{median}. Default \code{mean}}

\item{per_chr}{boolean; plot per chromosome}

\item{ignore_chr}{string; Chromsomes to ignore. If NULL, all chromosome will be used. Default \code{NULL}}

\item{ignore_samples}{list of strings; Samples to ignore.  If NULL, all samples will be used. Default \code{NULL}}

\item{n_col}{integer; number of columns. Passed to \code{facet_wrap}}

\item{n_row}{integer; number of rows. Passed to \code{facet_wrap}}

\item{pheno}{string; Col name of colData(m). Will be used as a factor to color different groups}

\item{verbose}{boolean; Flag for outputting function status messages. Default = TRUE}

\item{show_legend}{boolean; Display the legend on the plot}

\item{...}{Additional parameters to feed to scMethrix_theme()}
}
\value{
ggplot2 object
}
\description{
Plot descriptive statistics
}
\details{
plot descriptive statistics results from \code{\link{get_stats}}
}
\examples{
data('scMethrix_data')
plot_stats(scMethrix_data)
}
\seealso{
\code{\link{get_stats}}
}
