% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_plot.R
\name{plot_tsne}
\alias{plot_tsne}
\title{Plot tSNE results}
\usage{
plot_tsne(scm = NULL, col_anno = NULL, shape_anno = NULL, show_labels = FALSE)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{col_anno}{Column name of colData(m). Default NULL. Will be used as a factor to color different groups. Required \code{methrix} object}

\item{shape_anno}{Column name of colData(m). Default NULL. Will be used as a factor to shape different groups. Required \code{methrix} object}

\item{show_labels}{Show cell names on each data point. Default FLASE}
}
\value{
ggplot2 object
}
\description{
Plot tSNE results
}
\examples{
data('scMethrix_data')
scmpc = dim_red_scMethrix(scMethrix_data,type="tSNE")
plot_tsne(scmpc)
}
\seealso{
\code{\link[=tsne_scMethrix]{tsne_scMethrix()}} for dimensionality reduction
}
