% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_plot.R
\name{plot_umap}
\alias{plot_umap}
\title{Plot UMAP results}
\usage{
plot_umap(scm, col_anno = NULL, shape_anno = NULL, show_labels = FALSE)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the experiment containing the PCA}

\item{col_anno}{Column name of colData(m). Default NULL. Will be used as a factor to color different groups. Required \code{methrix} object}

\item{shape_anno}{Column name of colData(m). Default NULL. Will be used as a factor to shape different groups. Required \code{methrix} object}

\item{show_labels}{Show cell names on each data point. Default FLASE}
}
\value{
ggplot2 object
}
\description{
Plot UMAP results
}
\examples{
data('scMethrix_data')
}
