% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_plot.R
\name{plot_violin}
\alias{plot_violin}
\title{Violin Plot for \eqn{\beta}-Values}
\usage{
plot_violin(
  scm = NULL,
  assay = "score",
  n_cpgs = 25000,
  pheno = NULL,
  col_palette = "RdYlGn",
  show_legend = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{assay}{string; name of an existing assay. Default = "score"}

\item{n_cpgs}{integer; Use these many random CpGs for plotting. Default 25000. Set it to \code{NULL} to use all - which can be memory expensive. The seed will be set to \code{n_cpgs} for consistency.}

\item{pheno}{string; Col name of colData(m). Will be used as a factor to color different groups}

\item{col_palette}{string; Name of the RColorBrewer palette to use for plotting.}

\item{show_legend}{boolean; Display the legend on the plot}

\item{verbose}{boolean; Flag for outputting function status messages. Default = TRUE}

\item{...}{Additional parameters to feed to scMethrix_theme()}
}
\value{
ggplot2 object
}
\description{
Violin Plot for \eqn{\beta}-Values
}
\examples{
data('scMethrix_data')
plot_violin(scm = scMethrix_data)
}
