% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_plot.R
\name{prepare_plot_data}
\alias{prepare_plot_data}
\title{Format \code{\link{scMethrix}} matrix to long form data for plotting}
\usage{
prepare_plot_data(scm, ranges = NULL, n_cpgs = 25000, pheno = NULL)
}
\arguments{
\item{scm}{Input \code{\link{scMethrix}} object}

\item{ranges}{genomic regions to be summarized. Could be a data.table with 3 columns (chr, start, end) or a \code{GenomicRanges} object}

\item{n_cpgs}{Use these many random CpGs for plotting. Default 25000. Set it to \code{NULL} to use all - which can be memory expensive.}

\item{pheno}{Row name of colData(m). Will be used as a factor to color different groups in the violin plot.}
}
\value{
'Long' matrix for methylation
}
\description{
Format \code{\link{scMethrix}} matrix to long form data for plotting
}
