% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_beds.R
\name{read_bed_by_index}
\alias{read_bed_by_index}
\title{Parses BED files for methylation values using previously generated index genomic coordinates}
\usage{
read_bed_by_index(file, ref_cpgs, col_idx = NULL, zero_based = FALSE)
}
\arguments{
\item{file}{The BED file to parse}

\item{ref_cpgs}{The index of all unique coordinates from the input BED files}

\item{col_idx}{The column indexes for the input BED files}

\item{zero_based}{Whether the input data is 0 or 1 based}
}
\value{
data.table containing vector of all indexed methylation values for the input BED
}
\description{
Parses BED files for methylation values using previously generated index genomic coordinates
}
\details{
Creates an NA-based vector populated with methlylation values from the input BED file in the
respective indexed genomic coordinates
}
\examples{
\dontrun{
#Do Nothing
}
}
