% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_beds.R
\name{read_bed_by_index}
\alias{read_bed_by_index}
\title{Parses BED files for methylation values using previously generated index genomic coordinates}
\usage{
read_bed_by_index(
  files,
  ref_cpgs,
  col_list = NULL,
  zero_based = FALSE,
  strand_collapse = FALSE,
  fill = TRUE
)
}
\arguments{
\item{files}{string; file.paths of BED files to parse to parse}

\item{ref_cpgs}{data.table; The index of all unique coordinates from the input BED files}

\item{col_list}{string; The column index object for the input BED files}

\item{zero_based}{boolean; Whether the input data is 0 or 1 based}

\item{strand_collapse}{boolean; Default FALSE}

\item{fill}{boolean; Fill the output matrix to match the ref_cpgs. This must be used for HDF5 input formats to ensure consistent
spacing for the grid. It is optional for in-memory formats.}
}
\value{
data.table containing vector of all indexed methylation values for the input BED
}
\description{
Parses BED files for methylation values using previously generated index genomic coordinates
}
\details{
Creates an NA-based vector populated with methlylation values from the input BED file in the
respective indexed genomic coordinates
}
\examples{
\dontrun{
#Do Nothing
}
}
