% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_beds.R
\name{read_beds}
\alias{read_beds}
\title{Versatile BedGraph reader.}
\usage{
read_beds(
  files = NULL,
  ref_cpgs = NULL,
  colData = NULL,
  genome_name = "hg19",
  n_threads = 0,
  h5 = FALSE,
  h5_dir = NULL,
  h5_temp = NULL,
  desc = NULL,
  verbose = FALSE,
  zero_based = FALSE,
  reads = NULL,
  replace = FALSE
)
}
\arguments{
\item{files}{BED files containing methylation d}

\item{ref_cpgs}{BED files containing list of CpG sites. Must be zero-based genome.}

\item{genome_name}{Name of genome. Default hg19}

\item{n_threads}{number of threads to use. Default 1.
Be-careful - there is a linear increase in memory usage with number of threads. This option is does not work with Windows OS.}

\item{h5}{Should the coverage and methylation matrices be stored as 'HDF5Array'}

\item{h5_dir}{directory to store H5 based object}

\item{desc}{Description of the experiment}

\item{verbose}{Be little chatty ? Default TRUE.}

\item{stranded}{Default c}

\item{h5temp}{temporary directory to store hdf5}
}
\value{
An object of class \code{\link{scMethrix}}
}
\description{
Versatile BedGraph reader.
}
\details{
Reads BED files and generates methylation matrices.
Optionally arrays can be serialized as on-disk HDFS5 arrays.
}
\examples{
\dontrun{
bdg_files = list.files(path = system.file('extdata', package = 'methrix'),
pattern = '*\\\\.bedGraph\\\\.gz$', full.names = TRUE)
hg19_cpgs = methrix::extract_CPGs(ref_genome = 'BSgenome.Hsapiens.UCSC.hg19')
meth = methrix::read_bedgraphs( files = bdg_files, ref_cpgs = hg19_cpgs,
chr_idx = 1, start_idx = 2, M_idx = 3, U_idx = 4,
stranded = FALSE, zero_based = FALSE, collapse_strands = FALSE)
}

}
