% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_beds.R
\name{read_beds}
\alias{read_beds}
\title{Versatile BedGraph reader.}
\usage{
read_beds(
  files = NULL,
  ref_cpgs = NULL,
  colData = NULL,
  genome_name = "hg19",
  batch_size = 200,
  n_threads = 0,
  h5 = FALSE,
  h5_dir = NULL,
  h5_temp = NULL,
  desc = NULL,
  verbose = FALSE,
  zero_based = FALSE,
  reads = NULL,
  replace = FALSE,
  stranded = FALSE,
  coverage = FALSE,
  meth_idx = 4,
  cov_idx = NULL
)
}
\arguments{
\item{files}{BED files containing methylation d}

\item{ref_cpgs}{BED files containing list of CpG sites. Must be zero-based genome.}

\item{colData}{vector of input sample names}

\item{genome_name}{Name of genome. Default hg19}

\item{batch_size}{Max number of files to hold in memory at once}

\item{n_threads}{number of threads to use. Default 1.
Be-careful - there is a linear increase in memory usage with number of threads. This option is does not work with Windows OS.}

\item{h5}{Should the coverage and methylation matrices be stored as 'HDF5Array'}

\item{h5_dir}{directory to store H5 based object}

\item{h5_temp}{temporary directory to store hdf5}

\item{desc}{Description of the experiment}

\item{verbose}{flag to output messages or not.}

\item{zero_based}{Boolean flag for whether the input data is zero-based or not}

\item{reads}{Manual input of reads. Typically used for testing.}

\item{replace}{Boolean flag for whether to delete the contents of h5_dir before saving}

\item{stranded}{Default c}

\item{coverage}{flag for including a coverage matrix in the experiment}

\item{meth_idx}{The column index of the methylation value for the read}

\item{cov_idx}{The column index(es) of the read count. If not present, no coverage matrix is built}
}
\value{
An object of class \code{\link{scMethrix}}
}
\description{
Versatile BedGraph reader.
}
\details{
Reads BED files and generates methylation matrices.
Optionally arrays can be serialized as on-disk HDFS5 arrays.
}
