% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_beds.R
\name{read_hdf5_data}
\alias{read_hdf5_data}
\title{Writes values from input BED files into an in-disk \code{\link{HDF5Array}}}
\usage{
read_hdf5_data(
  files,
  ref_cpgs,
  col_list,
  n_threads = 0,
  h5_temp = NULL,
  zero_based = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{files}{The BED files to parse}

\item{ref_cpgs}{The index of all unique coordinates from the input BED files}

\item{col_list}{The column index object for the input BED files}

\item{n_threads}{The number of threads to use. 0 is the default thread with no cluster built.}

\item{h5_temp}{The file location to store the \code{\link{RealizationSink}} object}

\item{zero_based}{Boolean flag for whether the input data is zero-based or not}

\item{verbose}{flag to output messages or not.}
}
\value{
List of \code{\link{HDF5Array}}. 1 is methylation, 2 is coverage. If no cov_idx is specified, 2 will be NULL
}
\description{
Writes values from input BED files into an in-disk \code{\link{HDF5Array}}
}
\details{
Using the generated index for genomic coordinates, creates a NA-based dense matrtix of methylation
values for each BED file/sample. Each column contains the meth. values for a single sample.
}
\examples{
\dontrun{
#Do Nothing
}
}
