% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_beds.R
\name{read_hdf5_data}
\alias{read_hdf5_data}
\title{Writes methylation values from input BED files into an HDF5array}
\usage{
read_hdf5_data(
  files,
  ref_cpgs,
  n_threads = 0,
  h5_temp = NULL,
  zero_based = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{files}{The BED files to parse}

\item{ref_cpgs}{The index of all unique coordinates from the input BED files}

\item{n_threads}{The number of threads to use. 0 is the default thread with no cluster built.}

\item{h5_temp}{The file location to store the RealizationSink object}

\item{zero_based}{Boolean flag for whether the input data is zero-based or not}

\item{verbose}{flag to output messages or not.}
}
\value{
HDF5Array The methylation values for input BED files
}
\description{
Writes methylation values from input BED files into an HDF5array
}
\details{
Using the generated index for genomic coordinates, creates a NA-based dense matrtix of methylation
values for each BED file/sample. Each column contains the meth. values for a single sample.
}
