% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_beds.R
\name{read_index}
\alias{read_index}
\title{Parse BED files for unique genomic coordinates}
\usage{
read_index(
  files,
  col_list,
  n_threads = 0,
  zero_based = FALSE,
  batch_size = 200,
  verbose = TRUE
)
}
\arguments{
\item{files}{list of strings; file.paths of BED files}

\item{col_list}{string; The column index object for the input BED files}

\item{n_threads}{integer; number of threads to use. Default 1.
Be-careful - there is a linear increase in memory usage with number of threads. This option is does not work with Windows OS.}

\item{zero_based}{boolean; flag for whether the input data is zero-based or not}

\item{batch_size}{integer; Max number of files to hold in memory at once. Default 20}

\item{verbose}{boolean; flag to output messages or not.}
}
\value{
data.table containing all unique genomic coordinates
}
\description{
Parse BED files for unique genomic coordinates
}
\details{
Create list of unique genomic regions from input BED files. Populates a list of batch_size+1 with
the genomic coordinates from BED files, then runs \code{\link{unique}} when the list is full and keeps the running
results in the batch_size+1 position. Also indexes based on 'chr' and 'start' for later searching.
}
\examples{
\dontrun{
#Do Nothing
}
}
