% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_beds.R
\name{read_mem_data}
\alias{read_mem_data}
\title{Writes methylation values from input BED files into an in-memory \code{\link{RangedSummarizedExperiment}}}
\usage{
read_mem_data(
  files,
  ref_cpgs,
  batch_size = 200,
  n_threads = 0,
  zero_based = FALSE,
  meth_idx = 4,
  cov_idx = NULL,
  verbose = TRUE
)
}
\arguments{
\item{files}{The BED files to parse}

\item{ref_cpgs}{The index of all unique coordinates from the input BED files}

\item{batch_size}{The number of files to hold in memory at once}

\item{n_threads}{The number of threads to use. 0 is the default thread with no cluster built.}

\item{zero_based}{Boolean flag for whether the input data is zero-based or not}

\item{meth_idx}{The column index of the methylation value for the read}

\item{cov_idx}{The column index(es) of the read count}

\item{verbose}{flag to output messages or not.}
}
\value{
matrix of the methylation values for input BED files
}
\description{
Writes methylation values from input BED files into an in-memory \code{\link{RangedSummarizedExperiment}}
}
\details{
Using the generated index for genomic coordinates, creates a NA-based dense matrtix of methylation
values for each BED file/sample. Each column contains the meth. values for a single sample.
}
