% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_beds.R
\name{read_mem_data}
\alias{read_mem_data}
\title{Writes values from input BED files into an in-memory \code{\link{matrix}}}
\usage{
read_mem_data(
  files,
  ref_cpgs,
  col_list,
  batch_size = 200,
  n_threads = 0,
  zero_based = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{files}{The BED files to parse}

\item{ref_cpgs}{The index of all unique coordinates from the input BED files}

\item{col_list}{The column index object for the input BED files}

\item{batch_size}{The number of files to hold in memory at once}

\item{n_threads}{The number of threads to use. 0 is the default thread with no cluster built.}

\item{zero_based}{Boolean flag for whether the input data is zero-based or not}

\item{verbose}{flag to output messages or not.}
}
\value{
matrix of the methylation values for input BED files
}
\description{
Writes values from input BED files into an in-memory \code{\link{matrix}}
}
\details{
Using the generated index for genomic coordinates, creates a NA-based dense matrtix of methylation
values for each BED file/sample. Each column contains the meth. values for a single sample.
}
\examples{
\dontrun{
#Do Nothing
}
}
