% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_operations.R
\name{remove_assay}
\alias{remove_assay}
\title{Removes an assay from an \code{\link{scMethrix}} object.}
\usage{
remove_assay(scm, assay)
}
\arguments{
\item{scm}{A \code{\link{scMethrix}} object}

\item{assay}{The name of an assay that exists in the \code{\link{scMethrix}} object}
}
\value{
An \code{\link{scMethrix}} object
}
\description{
Removes an assay from an \code{\link{scMethrix}} object.
}
\examples{
data('scMethrix_data')
remove_assay(scMethrix_data,assay="counts")
}
