% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_operations.R
\name{remove_uncovered}
\alias{remove_uncovered}
\title{Remove loci that are uncovered across all samples}
\usage{
remove_uncovered(scm = NULL, verbose = TRUE)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{verbose}{boolean; Flag for outputting function status messages. Default = TRUE}
}
\value{
An object of class \code{\link{scMethrix}}
}
\description{
Remove loci that are uncovered across all samples
}
\details{
Takes \code{\link{scMethrix}} object and removes loci that are uncovered across all samples
}
\examples{
data('scMethrix_data')
# Remove uncovered CpGs after subsetting to a single sample
remove_uncovered(subset_scMethrix(scMethrix_data, samples = "C1", by="include"))
}
