% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_operations.R
\name{save_HDF5_scMethrix}
\alias{save_HDF5_scMethrix}
\title{Saves an HDF5 \code{\link{scMethrix}} object}
\usage{
save_HDF5_scMethrix(scm = NULL, h5_dir = NULL, replace = FALSE, ...)
}
\arguments{
\item{scm}{\code{\link{scMethrix}} object}

\item{h5_dir}{The directory to use. Created, if not existing. Default NULL}

\item{replace}{Should it overwrite the pre-existing data? FALSE by default.}

\item{...}{Parameters to pass to saveHDF5SummarizedExperiment}
}
\value{
Nothing
}
\description{
Saves an HDF5 \code{\link{scMethrix}} object
}
\details{
Takes \code{\link{scMethrix}} object and saves it in the specified directory
}
\examples{
data('scMethrix_data')
dir <- paste0(tempdir(),"/h5")
scm <- convert_scMethrix(scMethrix_data, h5_dir=dir)
save_HDF5_scMethrix(scm, h5_dir = dir, replace = TRUE)
}
