% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_data.R
\docType{data}
\name{scMethrix_data}
\alias{scMethrix_data}
\title{Simulated WGBS data}
\format{
A list containing two objects of class \code{'scMethrix'}, one in mem and one in H5 format.
}
\usage{
data('scMethrix_data')
}
\description{
This is a temporary small set of data representing CpGs on chr1 and chr2 in regions 1:20
}
\examples{
data('scMethrix_data')
scMethrix_data
}
\keyword{datasets}
