% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_data.R
\docType{data}
\name{scMethrix_data}
\alias{scMethrix_data}
\title{Single cell methylome of murine neuronal cells}
\format{
A object of class \code{'scMethrix'}, in-memory format.
}
\usage{
data('scMethrix_data')
}
\description{
This is a subset of the original converted to \code{scMethrix} format. See data_gen.R for subset details
}
\examples{
data('scMethrix_data')
scMethrix_data
}
\references{
Luo, C. et al. (2017). Single-cell methylomes identify neuronal subtypes and regulatory elements in mammalian cortex
}
\seealso{
\url{https://pubmed.ncbi.nlm.nih.gov/28798132/}

\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE97179}
}
\keyword{datasets}
