% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_plot.R
\name{scMethrix_theme}
\alias{scMethrix_theme}
\title{Theme for ggplot}
\usage{
scMethrix_theme(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{numeric; A sparsity proportion between 0 and 1. E.g. 0.1 replaces 10\% of the matrix with NA}

\item{base_family}{closure; The imputation methods to compare.}
}
\value{
list; data for the ggplot theme
}
\description{
Theme for ggplot
}
