% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_funcs.R
\name{split_granges}
\alias{split_granges}
\title{Chunks a Granges object by factor, percent or number}
\usage{
split_granges(gr, factor = NA, percent = NA, num = NA)
}
\arguments{
\item{gr}{The Granges object}

\item{factor}{The factor to which divide the chunks into}

\item{percent}{The percentage of the Granges to chunk}

\item{num}{The number of regions to include in each chunk}
}
\value{
GRangesList containing all the chunked Granges
}
\description{
Chunks a Granges object by factor, percent or number
}
\details{
Divides Granges into a list based on a chunking parameter
}
