% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_funcs.R
\name{split_vector}
\alias{split_vector}
\title{Splits a vector into subvectors}
\usage{
split_vector(v, n, by = c("chunk", "size"))
}
\arguments{
\item{v}{The vector to split}

\item{n}{The number to split by}

\item{by}{Whether to split by chunk or by size}
}
\value{
vector
}
\description{
Splits a vector into subvectors
}
\details{
Splits a vector into n chunks or chunks of n size.
}
\examples{
# Split vector into 4 sub vectors
split_vector(c(1,2,3,4,5,6,7,8),4,by="chunk")

# Split vector into sub-vectors with a size of 2
split_vector(c(1,2,3,4,5,6,7,8),2,by="size")
}
