% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_funcs.R
\name{split_vector}
\alias{split_vector}
\title{Splits a vector into list of vectors by \code{chunk} or \code{size}}
\usage{
split_vector(vec, chunks = NA, percent = NA, size = NA)
}
\arguments{
\item{vec}{vector; The vector to split}

\item{chunks}{integer; The number of desired sub-lists}

\item{percent}{integer; The maximum percentage of elements each sub-list should hold}

\item{size}{integer; The maximum size of each sub-list}
}
\value{
A list of sub-vectors
}
\description{
Splits a vector into list of vectors by \code{chunk} or \code{size}
}
\details{
Splits a vector into consistantly sized sub-lists. The sub-list size will always be decreasing based on list order.
}
\examples{
# Split vector into 4 sub vectors
split_vector(c(1,2,3,4,5,6,7,8),chunks=4)

# Split vector into sub-vectors with a size of 2
split_vector(c(1,2,3,4,5,6,7,8),size=2)

# Split vector into sub-vectors each with 25\% of the total elements
split_vector(c(1,2,3,4,5,6,7,8),percent=25)
}
