% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_funcs.R
\name{stop_time}
\alias{stop_time}
\title{Stops an internal stopwatch and outputs overall time}
\usage{
stop_time()
}
\value{
Returns formatted elapsed time since start_time
}
\description{
Stops an internal stopwatch and outputs overall time
}
\details{
Gets the stored elapsed proc.time() from initial start_time() to calculate overall runtime
}
