% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmethrix_operations.R
\name{subset_scMethrix}
\alias{subset_scMethrix}
\title{Subsets \code{\link{scMethrix}} object based on given conditions.}
\usage{
subset_scMethrix(
  m,
  regions = NULL,
  contigs = NULL,
  samples = NULL,
  verbose = TRUE
)
}
\arguments{
\item{m}{\code{\link{scMethrix}} object}

\item{regions}{genomic regions to subset by. Could be a data.table with 3 columns (chr, start, end) or a \code{GenomicRanges} object}

\item{contigs}{string of chromosome names to subset by}

\item{samples}{string of sample names to subset by}

\item{verbose}{flag to output messages or not}
}
\value{
An object of class \code{\link{scMethrix}}
}
\description{
Subsets \code{\link{scMethrix}} object based on given conditions.
}
\details{
Takes \code{\link{scMethrix}} object and filters CpGs based on coverage statistics
}
\examples{
data('scMethrix_data')
contigs <- c("chr1","chr3")
regions <- GenomicRanges::GRanges(seqnames = "chr1", ranges = IRanges(1,5)) 
samples <- c("df1","df3")

#Subset to chromosome 1
subset_scMethrix(scMethrix_data$mem, contigs = contigs)

#Subset to samples bed1 and bed3
subset_scMethrix(scMethrix_data$mem, samples = samples)

#Subset to samples bed1 and bed3, and chromosome 1
subset_scMethrix(scMethrix_data$mem, samples = samples, contigs = contigs)

#Subset to region "chr1:1-5"
subset_scMethrix(scMethrix_data$mem, regions = regions)
}
