% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_transforms.R
\name{transform_assay}
\alias{transform_assay}
\title{Transforms an assay in an \code{\link{scMethrix}} object.}
\usage{
transform_assay(
  scm,
  assay = NULL,
  new_assay = NULL,
  trans = NULL,
  h5_temp = NULL
)
}
\arguments{
\item{scm}{\code{\link{scMethrix}}; the single cell methylation experiment}

\item{assay}{string; name of an existing assay. Default = "score"}

\item{new_assay}{string; name for transformed assay. Default = "new_assay"}

\item{trans}{closure; The transformation function. Default = mean}

\item{h5_temp}{string; temporary directory to store the temporary HDF5 files}
}
\value{
An \code{\link{scMethrix}} object
}
\description{
Transforms an assay in an \code{\link{scMethrix}} object.
}
\details{
Uses the inputted function to transform an assay in the \code{\link{scMethrix}} object. It is typically used for The transform is applied column-wise to optimize how HDF5 files access sample data. If HDF5 objects are used, transform functions should be from \pkg{DelayedMatrixStats}.
}
\examples{
data('scMethrix_data')
transform_assay(scMethrix_data,assay="score",new_assay="plus1",trans=function(x){x+1})
}
