% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_transforms.R
\name{transform_assay}
\alias{transform_assay}
\title{Transforms an assay in an \code{\link{scMethrix}} object.}
\usage{
transform_assay(m, assay = NULL, name = NULL, trans = NULL, h5_temp = NULL)
}
\arguments{
\item{m}{A \code{\link{scMethrix}} object}

\item{assay}{String name of an existing assay}

\item{name}{String name of transformed assay}

\item{trans}{The transformation function}

\item{h5_temp}{temporary directory to store hdf5}
}
\value{
An \code{\link{scMethrix}} object
}
\description{
Transforms an assay in an \code{\link{scMethrix}} object.
}
\details{
Uses the inputted function to transform an assay in the \code{\link{scMethrix}} object
}
\examples{
data('scMethrix_data')
transform_assay(scMethrix_data,assay="score",name="plus1",trans=function(x){x+1})
}
