% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_dimensionality.R
\name{tsne_scMethrix}
\alias{tsne_scMethrix}
\title{Generates tSNE for scMethrix}
\usage{
tsne_scMethrix(
  scm,
  assay = "score",
  var = "top",
  top_var = 1000,
  perplexity = 30,
  verbose = FALSE,
  n_components = 2,
  ...
)
}
\arguments{
\item{scm}{Input \code{\link{scMethrix}} object}

\item{assay}{The assay to use. Default is 'score'}

\item{var}{Choose between random CpG sites ('rand') or most variable CpGs ('top').}

\item{top_var}{Number of variable CpGs to use. Default 1000 Set it to NULL to use all CpGs (which is not recommended due to memory requirements). This option is mutually exclusive with \code{ranges}.}

\item{perplexity}{numeric; Perplexity parameter (should not be bigger than 3 * perplexity < nrow(X) - 1, see details for interpretation)}

\item{verbose}{flag to output messages or not}

\item{n_components}{Number of components to use}

\item{...}{Other arguments that can be passed to Rtsne}
}
\value{
\code{\link{scMethrix}} object with reducedDim 'tSNE'
}
\description{
Generates tSNE for scMethrix
}
\details{
Does tSNE stuff
}
\examples{
data('scMethrix_data')
umap_scMethrix(scMethrix_data)
}
\seealso{
\code{\link[=plot_tsne]{plot_tsne()}} for plotting
}
