% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMethrix_dimensionality.R
\name{umap_scMethrix}
\alias{umap_scMethrix}
\title{Generates UMAP for scMethrix}
\usage{
umap_scMethrix(
  scm,
  assay = "score",
  var = "top",
  top_var = 1000,
  verbose = FALSE,
  n_neighbors = 15,
  n_components = 2,
  ...
)
}
\arguments{
\item{scm}{Input \code{\link{scMethrix}} object}

\item{assay}{The assay to use. Default is 'score'}

\item{var}{Choose between random CpG sites ('rand') or most variable CpGs ('top').}

\item{top_var}{Number of variable CpGs to use. Default 1000 Set it to NULL to use all CpGs (which is not recommended due to memory requirements). This option is mutually exclusive with \code{ranges}.}

\item{verbose}{flag to output messages or not}

\item{n_neighbors}{integer; number of nearest neighbors}

\item{n_components}{Number of components to use}

\item{...}{list of settings; values overwrite defaults from config;
see documentation of umap.default for details about available settings}
}
\value{
\code{\link{scMethrix}} object with reducedDim 'UMAP'
}
\description{
Generates UMAP for scMethrix
}
\details{
Does UMAP stuff
}
\examples{
data('scMethrix_data')
umap_scMethrix(scMethrix_data)
}
\seealso{
\code{\link[=plot_umap]{plot_umap()}} for plotting
}
